% Generated by roxygen2 (4.0.1): do not edit by hand
\name{agePrecision}
\alias{agePrec}
\alias{agePrecision}
\alias{plot.agePrec}
\alias{summary.agePrec}
\title{Compute measures of precision among sets of ages.}
\usage{
agePrecision(formula, data)

\method{summary}{agePrec}(object, what = c("precision", "difference",
  "absolute difference", "detail", "agreement"), percent = TRUE, digits = 4,
  ...)
}
\arguments{
\item{formula}{A formula of the form \code{~var1+var2+var3+...} where the \code{varX} generically represent the variables that contain the age assignments.  See details.}

\item{data}{A data.frame that minimally contains the age assignments in the variables identified in \code{formula}.}

\item{object}{An object saved from the \code{agePrecision} call (i.e., of class \code{agePrec}).}

\item{what}{A string that indicates what type of summary to print.  See details.}

\item{percent}{A logical that indicates whether the difference tableas (see details) should be represented as percentages (\code{TRUE}; default) or frequency of fish.}

\item{digits}{A single numeric that indicates the minimum number of digits to print when using \code{summary}.}

\item{\dots}{Additional arguments for methods.}
}
\value{
A list with the following items:
\itemize{
  \item detail A data.frame with all data given in \code{data} and intermediate calculations for each fish.  See details
  \item rawdiff A frequency table of fish by differences for each pair of ages.
  \item absdiff A frequency table of fish by absolute differences for each pair of ages.
  \item APE The mean average percent error.
  \item CV The mean coefficient of variation.
  \item n Number of fish in \code{data}.
  \item R Number of age assessments in \code{data}.
}

Nothing is returned by \code{summary}, but see details for what is printed.
}
\description{
Computes overall measures of precision for multiple age assignments made on the same individuals.  The age assignments may be from two or more readers of the same structure, one reader at two or more times, or two or more stuctures (e.g., scales, spines, otoliths).
}
\details{
The main function, \code{agePrecision}, requires a formula of the form \code{~var1+var2+var3+...} where the \code{varX} generically represent the variables that contain the multiple age assignments.  Alternatively, the formula may be of the form \code{var1~var2+var3+...} to allow for similar code as used in \code{\link{ageBias}}.  In the alternative formula the right-hand-side can have only one variable.

If \code{what="precision"} in \code{summary} then a summary table that contains the following items will be printed:
\itemize{
  \item n Number of fish in \code{data}.
  \item R Number of age assessments in \code{data}.
  \item CV The mean coefficient of variation.  See the fishR vignette for calculational details.
  \item APE The mean average percent error.  See the fishR vignette for calculational details.
  \item PercAgree The percentage of fish for which all age assignments perfectly agree.
}

If \code{what="difference"} in \code{summary} then a table that describes either the percentage (if \code{percent=TRUE}, default) or frequency of fish by the difference in paired age assignments.  This table has one row for each possible pair of age assignments.

If \code{what="absolute difference"} in \code{summary} then a table that describes either the percentage (if \code{percent=TRUE}, default) or frequency of fish by the absolute value of the difference in paired age assignments.  This table also has one row for each possible pair of age assignments.  The \dQuote{1} column, for example, represents age assignments that disagree by one year (in either direction).

If \code{what="detail"} in \code{summary} then a data frame of the original \code{data} along with the intermediate caculations of the average age, standard deviation of age, APE, and CV for each individual will be printed.  These details are generally only used to check or to understand calculations.
}
\section{fishR vignette}{
 \url{https://sites.google.com/site/fishrfiles/gnrl/AgeComparisons.pdf}
}
\examples{
## Example with just two age assignments
data(WhitefishLC)
ap1 <- agePrecision(~otolithC+scaleC,data=WhitefishLC)
summary(ap1)
summary(ap1,what="precision")
summary(ap1,what="difference")
summary(ap1,what="difference",percent=FALSE)
summary(ap1,what="absolute",percent=FALSE)

barplot(ap1$rawdiff,ylab="Frequency",xlab="Otolith - Scale Age")
summary(ap1,what="detail")
summary

## Example with three age assignments
ap2 <- agePrecision(~otolithC+finrayC+scaleC,data=WhitefishLC)
summary(ap2)
summary(ap2,what="precision")
summary(ap2,what="difference")
summary(ap2,what="difference",percent=FALSE)
summary(ap2,what="absolute",percent=FALSE)
summary(ap2,what="detail")
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
Beamish, R.J. and D.A. Fournier.  1981.  A method for comparing the precision of a set of age determinations.  Canadian Journal of Fisheries and Aquatic Sciences, 38:982-983. \url{http://www.pac.dfo-mpo.gc.ca/science/people-gens/beamish/PDF_files/compareagecjfas1981.pdf}

Campana, S.E.  1982.  Accuracy, precision and quality control in age determination, including a review of the use and abuse of age validation methods. Journal of Fish Biology, 59:197-242.  \url{http://www.denix.osd.mil/nr/crid/Coral_Reef_Iniative_Database/References_for_Reef_Assessment_files/Campana,\%202001.pdf}

Campana, S.E., M.C. Annand, and J.I. McMillan. 1995.  Graphical and statistical methods for determining the consistency of age determinations. Transactions of the American Fisheries Society, 124:131-138.  \url{http://www.bio.gc.ca/otoliths/documents/Campana\%20et\%20al\%201995\%20TAFS.pdf}

Chang, W.Y.B. 1982.  A statistical method for evaluating the reproducibility of age determination.  Canadian Journal of Fisheries and Aquatic Sciences, 39:1208-1210.  \url{http://www.nrcresearchpress.com/doi/abs/10.1139/f82-158}
}
\seealso{
See \code{\link{ageBias}} for computation of the full age agreement table, along with tests and plots of age bias.
}
\keyword{htest}
\keyword{manip}

