% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ALKAgeDist}
\alias{ALKAgeDist}
\title{Proportions-at-age from an Age-Length Key}
\usage{
ALKAgeDist(key, lenA.n, len.n)
}
\arguments{
\item{key}{A numeric matrix that contains the age-length key.  See details.}

\item{lenA.n}{A vector of sample sizes for each length interval in the \emph{aged sample}.}

\item{len.n}{A vector of sample sizes for each length interval in the \emph{complete sample} (i.e., all fish regardles of whether they were aged or not).}
}
\value{
A data.frame with as many rows as ages present in \code{key} and the following three variables:
\itemize{
  \item prop The prortion of fish at each age.
  \item se The SE for the proportion of fish at each age.
  \item cv The CV for the proportion of fish at each age.
 }
}
\description{
Uses the methods of Quinn and Deriso (1999) to compute the proportions-at-age (and the standard errors and coefficient of variation of those proportions) in a larger sample based on an age-length-key created from a subsample of ages through a two-stage random sampling design.
}
\details{
The age-length key sent in \code{key} must be constructed with length intervals as rows and age values as columns.  XXX
}
\section{fishR vignette}{
 none yet.
}

\section{Tests}{
 The results from this function perfectly match the results in Table 8.4 (left) of Quinn and Deriso when using \code{data(SnapperHG2)} from \pkg{FSAdata}.  The results also perfectly match the results from using \code{alkprop} in \pkg{fishmethods}.
}
\examples{
## None yet
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
Quinn, T. J. and R. B. Deriso. 1999. Quantitative Fish Dynamics. Oxford University Press, New York, New York. 542 pages
}
\seealso{
See \code{\link{ageKey}} and related functions for a completely different methodology.  See \code{alkprop} from \pkg{fishmethods} for the exact same methodology but with different inputs.
}
\keyword{manip}

