% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ALKMeanVar}
\alias{ALKMeanVar}
\title{Mean Values-at-age from an Age-Length Key}
\usage{
ALKMeanVar(formula, data, key, lenA.n, len.n)
}
\arguments{
\item{formula}{A formula of the form \code{var~length+age} where \dQuote{var} generically represents the variable to be summarized (e.g., length, weight, fecundity), \dQuote{length} generically represents the variable that contains the known length measurements, and \dQuote{age} generically represents the variable that contain the assessed ages.}

\item{data}{A data.frame that minimally contains the length measurements, assessed ages, and the variable to be summarized (i.e., this should be the aged sample) as given in \code{formula}.}

\item{key}{A numeric matrix that contains the age-length key.  See details.}

\item{lenA.n}{A vector of sample sizes for each length interval in the \emph{aged sample}.}

\item{len.n}{A vector of sample sizes for each length interval in the \emph{complete sample} (i.e., all fish regardles of whether they were aged or not).}
}
\value{
A data.frame with as many rows as ages present in \code{key} and the following three variables:
\itemize{
  \item mean The mean value at each age.
  \item se The SE for the mean value at each age.
  \item cv The CV for the mean value at each age.
 }
}
\description{
Uses the methods of Quinn and Deriso (1999) to compute the mean value-at-age (and the standard errors and coefficient of variation of those means) in a larger sample based on an age-length-key created from a subsample of ages through a two-stage random sampling design.  The mean values could be mean length-, weight-, or fecundity-at-age, for example.
}
\details{
The age-length key sent in \code{key} must be constructed with length intervals as rows and age values as columns.  XXX
}
\section{fishR vignette}{
 none yet.
}
\examples{
## None yet
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
Quinn, T. J. and R. B. Deriso. 1999. Quantitative Fish Dynamics. Oxford University Press, New York, New York. 542 pages
}
\seealso{
See \code{\link{ageKey}} and related functions for a completely different methodology.  See \code{\link{ALKAgeDist}} for a related method of determining the proportion of fish at each age.
}
\keyword{manip}

