% Generated by roxygen2 (4.0.1): do not edit by hand
\name{capHistSum}
\alias{capHistSum}
\title{Summarize capture histories in individual fish format.}
\usage{
capHistSum(df, cols2use = NULL)
}
\arguments{
\item{df}{A data.frame that contains the capture histories (and, perhaps, other information) in \dQuote{individual} fish format.  See details.}

\item{cols2use}{A numeric vector of columns that contain the capture histories.  See details.}
}
\value{
If the capture history data file represents only two samples, then a list with the following two components is returned.
 \itemize{
   \item \code{caphist} A vector summarizing the frequency of fish with each capture history.
   \item \code{sum} A data.frame that contains the number of marked fish from the first sample (\code{M}), the number of captured fish in the second sample (\code{n}), and the number of recaptured (i.e. previously marked) fish in the second sample (\code{m}).
 }

If the capture history data file represents more than two samples, then a list with the following four components is returned
 \itemize{
   \item \code{caphist} A vector summarizing the frequency of fish with each capture history.
   \item \code{sum} A data frame that contains the the number of captured fish in the ith sample (\code{n}), the number of recaptured (i.e. previously marked) fish in the ith sample (\code{m}), the number of marked fish returned to the population following the ith sample (\code{R}; this will equal \code{n} as the function currently does not handle mortalities); and the number of marked fish in the population prior to the ith sample (\code{M}).
   \item \code{methodB.top} A matrix that contains the top of the Method B table used for the Jolly-Seber method (i.e., a contingency table of capture sample (columns) and last seen sample (rows)).
   \item \code{methodB.bot} A data.frame that contains the bottom of the Method B table used for the Jolly-Seber method (i.e., the number of marked fish in the sample (\code{m}), the number of unmarked fish in the sample (\code{u}), the total number of fish in the sample (\code{n}), and the number of marked fish returned to the population following the sample (\code{R}).
 }
}
\description{
Use to summarize a capture history data file that is in the \dQuote{individual} fish format (see \code{\link{capHistConvert}} for a discussion of data file format types).  Summarized capture history results may be used in the Lincoln-Petersent, Schnabel, Schumacher-Eschmeyer, or Jolly-Seber methods for estimating population abundance (see \code{\link{mrClosed}} and \code{\link{mrOpen}}).
}
\details{
This function requires the capture history data file to be in the \dQuote{individual} fish format.  See \code{\link{capHistConvert}} for a description of this (and other) formats and for methods to convert from other formats to the \dQuote{individual} fish format.  In addition, this function requires only the capture history portion of the data file.  Thus, if \code{df} contains columns with non-capture history information (e.g., fish ID, length, location, etc.) then use \code{cols2use=} to identify which columns contain only the capture history information.  Columns to use can be identified by listing the column numbers (e.g., columns 2 through 7 could be included with \code{cols2use=2:7}).  In many instances it may be easier to identify columns to \emph{exclude} which can be done by preceding the column number by a negative sign (e.g., columns 1 through 3 are excluded with \code{cols2use=-(1:3)}).

The object returned from this function can be used directly in \code{\link{mrClosed}} and \code{\link{mrOpen}}.  See examples of this functionality on the help pages for those functions.
}
\note{
This function assumes that all unmarked captured fish are marked and returned to the population (i.e., no losses at the time of marking are allowed).
}
\section{fishR vignette}{
 \url{https://sites.google.com/site/fishrfiles/gnrl/MRClosed.pdf}, \url{https://sites.google.com/site/fishrfiles/gnrl/MROpen.pdf}
}
\examples{
# data.frame with IDs in the first column
data(PikeNYPartial1)
head(PikeNYPartial1)
capHistSum(PikeNYPartial1,cols2use=-1)

# An examle with only two sample events (for demonstration only)
capHistSum(PikeNYPartial1,cols2use=-c(1,4:5))
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
See \code{\link{capHistConvert}} for a descriptions of capture history data file formats and how to convert between them.  See \code{\link{mrClosed}} and \code{\link{mrOpen}} for how to estimate abundance from the summarized capture history information.
}
\keyword{manip}

