% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/changesPos.R
\name{changesPos}
\alias{changesPos}
\title{Find positions in a vector that have a different value from the previous position.}
\usage{
changesPos(x, include.first = TRUE)
}
\arguments{
\item{x}{A vector of sorted (generally) values.}

\item{include.first}{A logical that indicates whether the returned vector of positions should have a ``1'' at the beginning (i.e., by definition the first position is the first position that is different than the previous value).}
}
\value{
A vector of positions in \code{x} where the value differs from the value in the previous position.
}
\description{
Find positions in a vector that have a different value from the previous position.
}
\examples{
x <- rep(LETTERS[1:6],times=c(5,4,1,1,2,4))
data.frame(pos=1:length(x),x)  # for demonstration only
changesPos(x)
v <- rep(10:17,times=c(4,1,1,2,4,2,1,1))
data.frame(pos=1:length(v),v)  # for demonstration only
changesPos(v)
## An uninteresting example -- i.e., vector not sorted
z <- sample(10:17,20,replace=TRUE)
data.frame(pos=1:length(z),z)  # for demonstration only
changesPos(z)
z1 <- z[order(z)]
data.frame(pos=1:length(z1),z1)  # for demonstration only
changesPos(z1)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{manip}

