% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Ecoli.R
\docType{data}
\name{Ecoli}
\alias{Ecoli}
\title{Population growth of Escherichia coli.}
\format{A data frame with 8 observations on the following 2 variables:
 \describe{
   \item{days}{Elapsed duration of the experiment}
   \item{cells}{Number of cells in the population}
 }}
\source{
McKendrick, A.G. and M. Kesava Pai. 1911. The Rate of Multiplication of Micro-Organisms: a Mathematical Study. Proceedings of the Royal Society of Edinburgh. 31: 649-655.
}
\description{
The number of \emph{Escherichia coli} cells versus time.
}
\concept{
'Nonlinear Model'
}
\keyword{datasets}

