% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lrt}
\alias{extraSS}
\alias{lrt}
\alias{print.extraTest}
\title{Likelihood ratio and extra sum-of-squares tests.}
\usage{
lrt(sim, ..., com)

extraSS(sim, ..., com)

\method{print}{extraTest}(x, ...)
}
\arguments{
\item{sim}{The results of one \code{lm} or \code{nls} model, for example, that is a nested subset of the model in \code{com=}.}

\item{com}{The results of one \code{lm} or \code{nls} model, for example, that the models in \code{sim=} and \code{\dots} are a subset of.}

\item{x}{An object from \code{lrt()} or \code{extraSS()}.}

\item{\dots}{More model results that are nested subsets of the model in \code{com=}.}
}
\value{
The main function returns a matrix with as many rows as model comparisons and columns of the following types:
 \itemize{
   \item \code{DfO} The degrees-of-freedom from the subset (more simple) model.
   \item \code{RSSO}, \code{logLikO} The residual sum-of-squares (from \code{extraSS}) or log-likelihood (from \code{lrt}) from the subset (more simple) model.
   \item \code{DfA} The degrees-of-freedom from the \code{com=} model.
   \item \code{RSSA}, \code{logLikA} The residual sum-of-squares (from \code{extraSS}) or log-likelihood (from \code{lrt}) from the \code{com=} model.
   \item \code{Df} The difference in degrees-of-freedom between the two models.
   \item \code{SS}, \code{logLik} The difference in residual sum-of-squares (from \code{extraSS}) or log-likelihood (from \code{lrt}) between the two models.
   \item \code{F}, \code{Chisq} The corresponding F- (from \code{extraSS}) or Chi-square (from \code{lrt}) test statistic.
   \item \code{Pr(>F)"}, \code{Pr(>Chisq)"} The corresponding p-value.
 }
}
\description{
Likelihood ratio and extra sum-of-squares tests with multiple models nested at the same level below one common model.
}
\details{
\code{anova()} and \code{lrtest()} (from \pkg{lmtest}) provide simple methods for conducting extra sum-of-squares or likelihood ratio tests when one model is nested within another model or when there are several layers of simple models all nested within each other.  However, to compare several models that are nested at the same level with one common more complex model then \code{anova()} and \code{lrtest()} need to be repeated for each comparison.  This repetition can be eliminated with \code{lapply()} but then the output is voluminous.  This function is designed to remove the repetitiveness and to provide output that is compact and easy to read.
}
\note{
This is experimental at this point.  It seems to work fine for \code{nls} models but has not been tested thoroughly or with other types of models.
}
\examples{
## Nothing yet:
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{htest}

