% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fact2num}
\alias{fact2num}
\title{Converts "numeric" factor levels to numeric values.}
\usage{
fact2num(object)
}
\arguments{
\item{object}{A vector with \dQuote{numeric} factor levels to be converted to numeric values.}
}
\value{
A numeric vector.
}
\description{
Converts \dQuote{numeric} factor levels to numeric values.
}
\examples{
junk <- factor(c(1,7,2,4,3,10))
str(junk)
junk2 <- fact2num(junk)
str(junk2)

## Below not run because it will result in an error -- levels are not 'numeric'
\dontrun{
bad <- factor(c("A","B","C"))
bad2 <- fact2num(bad)
}
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{manip}

