% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGrowthStarts.R
\name{findGrowStarts}
\alias{findGrowStarts}
\title{Find reasonable starting values for common fish growth functions.}
\usage{
findGrowStarts(
  formula,
  data,
  type = c("von Bertalanffy", "Gompertz", "logistic", "Richards"),
  param = 1,
  constvals = NULL,
  fixed = NULL,
  plot = FALSE
)
}
\arguments{
\item{formula}{A formula of the form \code{len~age}.}

\item{data}{A data frame that contains the variables in \code{formula}.}

\item{type}{A string for the model type to use (von Bertalanffy, Gompertz, logistic, Richards).}

\item{param}{A numeric that indicates the specific parameterization of the model in \code{type} to use. See details.}

\item{constvals}{A NAMED vector or list of constant values (either lengths or ages) to be used in some of the von Bertalanffy parameterizations. See details.}

\item{fixed}{A NAMED vector or list that contains user-defined (i.e., fixed rather than automatically generated) starting values for one or more parameters. See details.}

\item{plot}{A logical that indicates whether a plot of the data with the superimposed model fit at the starting values should be created. This plot is for diagnostic purposes and, thus, cannot be modified in this function.}
}
\value{
A list that contains reasonable starting values. Note that the parameters will be listed with the same names in the same order as listed in \code{\link{makeGrowthFun}}.
}
\description{
Finds reasonable starting values for the parameters in a specific parameterization of common growth functions (von Bertalanffy, Gompertz, logistic, Richards).
}
\details{
This function attempts to find reasonable starting values for a variety of parameterizations of common functions used to model fish growth (von Bertalanffy, Gompertz, logistic, Richards). The starting values tend to work well in \code{nls} and related non-linear modeling functions, but there is no guarantee that they are the \sQuote{best} starting values (especially if the model is not appropriate for the data). One should perform sensitivity analyses to determine the impact of different starting values on the final model results.

BLAH-BLAH-BLAH
}
\note{
BLAH-BLAH-BLAH
}
\section{IFAR Chapter}{
 12-Individual Growth.
}

\examples{
##----- Make fake data
# Setup ages, sample sizes (general reduction in numbers with
#   increasing age), and additive SD to model
t <- 0:15
n <- c(5,10,40,35,25,12,10,10,8,6,5,3,3,3,2,2)
sd <- 15
# Expand ages and put in a data.frame
df <- data.frame(age=rep(t,n))
# Add lengths from 1st parameterization of each model type with
#   random error for individuals
vb1 <- makeGrowthFun(type="von Bertalanffy")
df$tlv <- round(vb1(df$age,Linf=450,K=0.3,t0=-0.5)+rnorm(sum(n),0,sd),0)
g1 <- makeGrowthFun(type="Gompertz")
df$tlg <- round(g1(df$age,Linf=450,a=1,g=0.3)+rnorm(sum(n),0,sd),0)
l1 <- makeGrowthFun(type="logistic")
df$tll <- round(l1(df$age,Linf=450,gninf=0.3,ti=2)+rnorm(sum(n),0,sd),0)
r1 <- makeGrowthFun(type="Richards")
df$tlr <- round(r1(df$age,Linf=450,ti=2,k=0.5,b1=1.5)+rnorm(sum(n),0,sd),0)
# brief view of data
head(df)

##----- Example starting values for 1st parameterization of each type
findGrowStarts(tlv~age,data=df,type="von Bertalanffy")
findGrowStarts(tlg~age,data=df,type="Gompertz")
findGrowStarts(tll~age,data=df,type="logistic")
findGrowStarts(tlr~age,data=df,type="Richards")

}
\seealso{
See \code{\link{growthFunShow}} to make functions that use these starting values and \code{\link{showGrowFun}} to display the equations used in \pkg{FSA}. See \code{\link{nlsTracePlot}} for help troubleshooting nonlinear models that don't converge.
}
\author{
Derek H. Ogle, \email{DerekOgle51@gmail.com}
}
\keyword{manip}
