% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSA-defunct.R
\name{FSA-defunct}
\alias{FSA-defunct}
\alias{bootCase}
\alias{chooseColors}
\alias{compIntercepts}
\alias{compSlopes}
\alias{diags}
\alias{filterD}
\alias{fitPlot}
\alias{fsaNews}
\alias{hoCoef}
\alias{mapvalues}
\alias{plotBinResp}
\alias{residPlot}
\alias{Subset}
\alias{Schnute}
\title{DEFUNCT functions.}
\usage{
bootCase(...)

chooseColors(...)

compIntercepts(...)

compSlopes(...)

diags(...)

filterD(...)

fitPlot(...)

fsaNews(...)

hoCoef(...)

mapvalues(...)

plotBinResp(...)

residPlot(...)

Subset(...)

Schnute(...)
}
\arguments{
\item{\dots}{Additional arguments to pass through.}
}
\description{
These functions were once part of FSA but have now been removed.
}
