% Generated by roxygen2 (4.0.1): do not edit by hand
\name{.onAttach}
\alias{.onAttach}
\alias{iCILabel}
\alias{iCheckStartcatW}
\alias{iGetVarFromFormula}
\alias{iHndlFormula}
\alias{iHndlMultWhat}
\alias{iLegendHelp}
\alias{iListSpecies}
\alias{iMakeColor}
\alias{iTypeoflm}
\title{Internal functions used in FSA.}
\usage{
.onAttach(lib, pkg, ...)
}
\description{
Internal functions used in FSA
}
\details{
Take note of the following uses:
 \itemize{
   \item \code{iCheckStartcatW} used in \code{\link{lencat}} and \code{\link{lenFreqExpand}}.
   \item \code{iCILabel} is used in \code{\link{binCI}}, \code{\link{bootCase}}, \code{\link{catchCurve}}, \code{\link{chapmanRobson}}, \code{\link{confint.nlsBoot}}, \code{\link{depletion}}, \code{\link{hyperCI}}, \code{\link{mrClosed}}, \code{\link{poiCI}}, \code{\link{predict.nlsBoot}}, \code{\link{removal}}.
   \item \code{iGetVarFromFormula} is used in \code{\link{ageKey}}, \code{\link{lencat}}, \code{\link{psdCalc}}, \code{\link{psdDataPrep}}, \code{\link{psdPlot}}, \code{\link{recodeSpecies}}, \code{\link{wrAdd}}, and \code{\link{wrDataPrep}}.
   \item \code{iHndlFormula} is used in \code{\link{ageBias}}, \code{\link{agePrecision}}, \code{\link{growthModelSim}}, \code{\link{vbStarts}},  \code{\link{walfordPlot}}, and \code{\link{chapmanPlot}}.
   \item \code{iHndlMultWhat} is used in \code{\link{ageBias}} and \code{\link{agePrecision}}.
   \item \code{iLegendHelp} is used in \code{\link{fitPlot}} and \code{\link{residPlot}}.
   \item \code{iListSpecies} is used in \code{\link{psdVal}} and \code{\link{wsVal}}.
   \item \code{iMakeColor} is used in \code{\link{plotBinResp}} and \code{\link{tictactoe}}.
   \item \code{iTypeoflm} is used in \code{\link{fitPlot}} and \code{\link{residPlot}}.
 }
}
\keyword{internal}

