% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fsaNews}
\alias{FSANews}
\alias{fsaNews}
\title{Read news and changes in the 'FSA' package.}
\usage{
fsaNews()
}
\value{
None.
}
\description{
Read news and changes in the \sQuote{FSA} package.
}
\examples{
## Not run because it will open an external browser.
\dontrun{
fsaNews()
FSANews()
}
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{manip}

