% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{Garvey4a}
\alias{Garvey4a}
\title{Desity of shad captured in offshore zones and sunfish seined in inshore areas
of four Ohio reservoirs across 22 years.}
\format{A data frame of 22 observations on the following 2 variables:
\describe{
\item{shad}{Number of shad per m^2.}
\item{sunfish}{Number of sunfish per m^2.}
}}
\source{
From Figure 4a in Garvey, J.E., E.A. Marschall, and R.A. Wright. 1998.
From star charts to stoneflies: detecting relationships in continuous bivariate
data. Ecology 79:442-447.
}
\description{
Desity of shad captured in offshore zones and sunfish seined in inshore areas
of four Ohio reservoirs across 22 years.
}
\examples{
data(Garvey4a)
str(Garvey4a)
head(Garvey4a)
plot(sunfish~shad,data=Garvey4a,pch=16)
}
\keyword{datasets}

