% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompertzStarts.R
\name{GompertzStarts}
\alias{GompertzStarts}
\title{Find reasonable starting values for a Gompertz growth function.}
\usage{
GompertzStarts(
  formula,
  data = NULL,
  param = c("Ricker1", "Ricker2", "Ricker3", "QuinnDeriso1", "QuinnDeriso2",
    "QuinnDeriso3", "QD1", "QD2", "QD3", "Original", "original"),
  type = param,
  fixed = NULL,
  plot = FALSE,
  col.mdl = "gray70",
  lwd.mdl = 3,
  lty.mdl = 1,
  cex.main = 0.9,
  col.main = "red",
  ...
)
}
\arguments{
\item{formula}{A formula of the form \code{len~age}.}

\item{data}{A data frame that contains the variables in \code{formula}.}

\item{type, param}{A string that indicates the parameterization of the von Bertalanffy model.}

\item{fixed}{A named list that contains user-defined rather than automatically generated (i.e., fixed) starting values for one or more parameters. See details.}

\item{plot}{A logical that indicates whether a plot of the data with the superimposed model fit at the starting values should be created.}

\item{col.mdl}{A color for the model when \code{plot=TRUE}.}

\item{lwd.mdl}{A line width for the model when \code{plot=TRUE}.}

\item{lty.mdl}{A line type for the model when \code{plot=TRUE}.}

\item{cex.main}{A character expansion value for the main title when \code{plot=TRUE}.}

\item{col.main}{A color for the main title when \code{plot=TRUE}.}

\item{\dots}{Further arguments passed to the methods.}
}
\value{
A list that contains reasonable starting values. Note that the parameters will be listed in the same order and with the same names as listed in \code{\link{GompertzFuns}}.
}
\description{
Finds reasonable starting values for the parameters in a specific parameterization of the Gompertz growth function.
}
\details{
This function attempts to find reasonable starting values for a variety of parameterizations of the Gompertz growth function. There is no guarantee that these starting values are the \sQuote{best} starting values. One should use them with caution and should perform sensitivity analyses to determine the impact of different starting values on the final model results.

BLAH-BLAH-BLAH
}
\note{
The \sQuote{Ricker2} and \sQuote{QuinnDeriso1} are synonymous, as are \sQuote{Ricker3} and \sQuote{QuinnDeriso2}.
}
\section{IFAR Chapter}{
 12-Individual Growth.
}

\examples{
# Make some fake data using the original parameterization
gompO <- GompertzFuns("original")
# setup ages, sample sizes (general reduction in numbers with
# increasing age), and additive SD to model
t <- 0:15
n <- c(5,10,40,35,25,12,10,10,8,6,5,3,3,3,2,2)
sd <- 15
# expand ages
ages <- rep(t,n)
# get lengths from gompertz and a random error for individuals
lens <- gompO(ages,Linf=450,a=1,gi=0.3)+rnorm(length(ages),0,sd)
# put together as a data.frame
df <- data.frame(age=ages,tl=round(lens,0))

GompertzStarts(tl~age,data=df)
GompertzStarts(tl~age,data=df,plot=TRUE)

GompertzStarts(tl~age,data=df,param="Ricker2",plot=TRUE)
GompertzStarts(tl~age,data=df,param="Ricker3",plot=TRUE)
GompertzStarts(tl~age,data=df,param="QuinnDeriso3",plot=TRUE)
GompertzStarts(tl~age,data=df,param="Original",plot=TRUE)

}
\seealso{
See \code{\link{growthFunShow}} to display the equations for the parameterizations used in \pkg{FSA} and \code{\link{GompertzFuns}} for functions that represent the von Bertalanffy parameterizations. See \code{\link{nlsTracePlot}} for help troubleshooting nonlinear models that don't converge.
}
\author{
Derek H. Ogle, \email{DerekOgle51@gmail.com}
}
\keyword{manip}
