% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gompModels.R
\name{gompModels}
\alias{gompModels}
\title{Show the Gompertz function formulae implemented in the FSA package.}
\usage{
gompModels(...)
}
\arguments{
\item{\dots}{Additional arguments for \code{\link{plot}}.  Generally not needed}
}
\value{
A graphic that uses \code{\link{plotmath}} to show the model formulae in a pretty format.
}
\description{
Show the Gompertz function formulae implemented in \code{\link{gompFuns}}.
}
\section{IFAR Chapter}{
 None specifically, but \href{https://fishr.wordpress.com/books/ifar/}{9-Individual Growth} is related.
}
\examples{
\dontrun{windows(5,5)}
gompModels()
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
See references in \code{\link{gompFuns}}.
}
\seealso{
See \code{\link{gompFuns}} for functions that represent the Gompertz parameterizations.
}
\keyword{hplot}
\keyword{manip}

