% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growthLogistic.R
\name{growthLogistic}
\alias{growthLogistic}
\alias{logisticFuns}
\alias{logisticModels}
\title{Creates a function for a specific parameterization of the logistic growth function.}
\usage{
logisticFuns(type = c("CJ1", "CJ2", "Karkach", "HaddonI"), simple = FALSE,
  msg = FALSE)

logisticModels(type = c("size", "tagging"), cex = 1.25, ...)
}
\arguments{
\item{type}{A string that indicates the parameterization of the logistic growth function.}

\item{simple}{A logical that indicates whether the user should be allowed to send all parameter values in the first parameter argument (\code{=FALSE}; default) or whether all individual parameters must be specified (\code{=TRUE}).}

\item{msg}{A logical that indicates whether a message about the model and parameter definitions should be output (\code{=TRUE}) or not (\code{=FALSE}; default).}

\item{cex}{A single numeric expansion value for use with \code{logisticModels}.}

\item{\dots}{Not implemented.}
}
\value{
\code{logisticFuns} returns a function that can be used to predict fish length or weight given a vector of ages and values for the function parameters.  The result should be saved to an object that can then be used as a function name.  When the resulting function is used, the parameters are ordered as shown when the definitions of the parameters are printed after the function is called (if \code{msg=TRUE}).  If \code{simple=FALSE}, then the values for all parameters may be included as a vector in the first parameter argument.  If \code{simple=TRUE}, then all parameters must be declared individually.  The resulting function is somewhat easier to read when \code{simple=TRUE}.

\code{logisticModels} returns a graphic that uses \code{\link{plotmath}} to show the model formulae in a pretty format.
}
\description{
Creates a function for a specific parameterization of the logistic growth function.  Use \code{logisticModels()} to see the equations for each model.
}
\note{
Within FSA, L0 is the mean length at age 0, Linf is the mean asymptotic length, ti is the age at the inflection point, gninf is the instantaneous growth rate at negative infinity.
}
\section{IFAR Chapter}{
 None specifically, but \href{https://fishr.wordpress.com/books/ifar/}{9-Individual Growth} is related.
}
\examples{
## See the formulae
\dontrun{windows(5,5)}
logisticModels()

## Simple Examples
( log1 <- logisticFuns() )               # First Campana-Jones parameterization
ages <- 0:15
plot(log1(ages,Linf=800,gninf=0.5,ti=5)~ages,type="b",pch=19)

( log2 <- logisticFuns("CJ2") )          # Second Campana-Jones parameterization
plot(log2(ages,Linf=800,gninf=0.5,a=10)~ages,type="b",pch=19)

( log2c <- logisticFuns("CJ2",simple=TRUE) ) # compare to log2

( log3 <- logisticFuns("Karkach") )       # Karkach parametrization
plot(log3(ages,L0=10,Linf=800,gninf=0.5)~ages,type="b",pch=19)

( log4 <- logisticFuns("HaddonI") )       # Haddon's inverse logistic

#######################################################################################
## Examples of fitting Logistic models
##   After the last example a plot is constructed with three lines on top of each
##   other illustrating that the parameterizations all produce the same fitted values.
# Make some fake data using the original parameterization

gompO <- GompertzFuns("original")
# setup ages, sample sizes (general reduction in numbers with
# increasing age), and additive SD to model
t <- 1:15
n <- c(10,40,35,25,12,10,10,8,6,5,3,3,3,2,2)
sd <- 15
# expand ages
ages <- rep(t,n)
# get lengths from gompertz and a random error for individuals
lens <- gompO(ages,Linf=450,a=1,gi=0.3)+rnorm(length(ages),0,sd)
# put together as a data.frame
df <- data.frame(age=ages,len=round(lens,0))

# Fit first Campana-Jones parameterization
fit1 <- nls(len~log1(age,Linf,gninf,ti),data=df,start=list(Linf=450,gninf=0.45,ti=4))
summary(fit1,correlation=TRUE)
plot(len~age,data=df,pch=19,col=rgb(0,0,0,1/5))
curve(log1(x,Linf=coef(fit1)),from=0,to=15,col="red",lwd=10,add=TRUE)

# Fit second Campana-Jones parameterization
fit2 <- nls(len~log2(age,Linf,gninf,a),data=df,start=list(Linf=450,gninf=0.45,a=7))
summary(fit2,correlation=TRUE)
curve(log2(x,Linf=coef(fit2)),from=0,to=15,col="blue",lwd=5,add=TRUE)

# Fit Karkach parameterization (using simple=TRUE model)
log3 <- logisticFuns("Karkach",simple=TRUE)
fit3 <- nls(len~log3(age,Linf,L0,gninf),data=df,start=list(Linf=450,L0=30,gninf=0.45))
summary(fit3,correlation=TRUE)
curve(log3(x,Linf=coef(fit3)[1],L0=coef(fit3)[2],gninf=coef(fit3)[3]),
      from=0,to=15,col="green",lwd=2,add=TRUE)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}, thanks to Gabor Grothendieck for a hint about using \code{get()}.
}
\references{
Campana, S.E. and C.M. Jones.  1992.  \href{http://www.dfo-mpo.gc.ca/Library/141734.pdf}{Analysis of otolith microstructure data}.  Pages 73-100 In D.K. Stevenson and S.E. Campana, editors.  Otolith microstructure examination and analysis.  Canadian Special Publication of Fisheries and Aquatic Sciences 117.

Haddon, M., C. Mundy, and D. Tarbath. 2008.  \href{http://aquaticcommons.org/8857/1/haddon_Fish_Bull_2008.pdf}{Using an inverse-logistic model to describe growth increments of blacklip abalone (\emph{Haliotis rubra}) in Tasmania}.  Fishery Bulletin 106:58-71.

Karkach, A. S.  2006.  \href{http://www.demographic-research.org/volumes/vol15/12/15-12.pdf}{Trajectories and models of individual growth}.  Demographic Research 15:347-400.
}
\seealso{
See \code{\link{vbFuns}}, \code{\link{GompertzFuns}}, \code{\link{logisticFuns}}, and \code{\link{Schnute}} for similar functionality for other models.
}
\keyword{manip}

