% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/growthModels.R
\name{growthModels}
\alias{growthModels}
\alias{vbModels}
\title{Show the growth model formulas implemented in the FSA package.}
\usage{
growthModels(...)
}
\arguments{
\item{\dots}{Additional arguments for \code{plot}.  Generally not needed}
}
\value{
A graphic that uses \code{plotmath} to show the model formulas in a pretty format. \code{vbModels()} shows just the von Bertalanffy models.
}
\description{
Show the growth model formulas implemented in \code{\link{vbFuns}}, \code{\link{vbStarts}}, and \code{growthModelSim}.
}
\examples{
\dontrun{windows(8,5)}
growthModels()
vbModels()
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
See \code{\link{vbFuns}} for functions that represent the von Bertalanffy parameterizations and \code{\link{vbStarts}} and \code{\link{growthModelSim}} for methods to find starting values.
}
\keyword{hplot}
\keyword{manip}

