% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growthRadPlot.R
\name{growthRadPlot}
\alias{growthRadPlot}
\title{Constructs skeleton plots of radial or proportional radial measurements made at annuli on fish calcified structures.}
\usage{
growthRadPlot(df, indivs, in.pre = "rad", in.var, radcap = "radcap",
  agevar = "age", proportions = FALSE, numperpg = 20,
  xlab = ifelse(proportions, "Radial Measurements",
  "Proportions of Total Radius"), lwd = 2, col = "black", pch = 124,
  ymar = 5, ...)
}
\arguments{
\item{df}{A data frame that contains the radial growth measurement data in one-fish-per-line format.}

\item{indivs}{A string that contains the variable name or a vector of strings that contains the variable names that identify individual fish in the data frame.}

\item{in.pre}{A string that indicates the prefix for all variable names that contains the radial growth measurement data in the input data frame.}

\item{in.var}{A vector of column numbers or variable names that contain the radial growth measurement data in the input data frame.}

\item{radcap}{A string that contains the variable name that holds the total structure radius at the time of capture.}

\item{agevar}{A string that contains the variable name that holds the assigned ages.}

\item{proportions}{A logical that indicates whether raw radial measurements (\code{=FALSE}, default) or proportional (of the total radius) measurements are plotted.}

\item{numperpg}{A numeric that indicates the number of individuals to plot per page.}

\item{xlab}{A string for labeling the x-axis.}

\item{lwd}{A numeric that indicates the width of the lines plotted.}

\item{col}{A string that indicates the color to use for the plotted lines.}

\item{pch}{A numeric that indicates the plotting character used to mark the annuli.  The default (\code{=124}) is a vertical line.}

\item{ymar}{A numeric that indicates how wide the y-axis margin should be for labelling the individuals.}

\item{\dots}{Optional arguments sent to the \code{lines()} or \code{points()}.}
}
\value{
None.  However, a plot is produced.
}
\description{
Constructs skeleton plots of radial or proportional radial measurements made at annuli on fish calcified structures.  These plots can be used to identify clearly incorrect radial measurements.
}
\details{
This function constructs a plot of horizonal lines for each individual with the annular increments marked with vertical lines.  This is simply an exploratory graph to allow visual comparison of patterns among individuals.

This function requires one-fish-per-line radial measurement data.  See \code{\link{gReshape}} and \code{\link{gConvert}} if your data is one-increment-per-line or increment measurements.  This function also assumes that the data frame has the total radius-at-capture and the fish's age.  See \code{\link{addRadCap}} to add the total radius at capture variable (if your data was increments).
}
\examples{
## ONLY RUN IN INTERACTIVE MODE
if (interactive()) {

data(SMBassWB)
head(SMBassWB)     # to see column names & some data
growthRadPlot(SMBassWB,c("yearcap","gear","fish"),
              in.pre="anu",radcap="radcap",agevar="agecap",ymar=6)

} ## END IF INTERACTIVE MODE
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
\code{\link{gReshape}}, \code{\link{gConvert}}
}
\keyword{manip}

