% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growthRichards.R
\name{growthRichards}
\alias{RichardsFuns}
\alias{RichardsModels}
\alias{growthRichards}
\title{Creates a function for a specific parameterization of the Richards growth function.}
\usage{
RichardsFuns(param = 1, simple = FALSE, msg = FALSE)

RichardsModels(...)
}
\arguments{
\item{param}{A single numeric that indicates the parameterization of the Richards growth function.}

\item{simple}{A logical that indicates whether the user should be allowed to send all parameter values in the first parameter argument (\code{=FALSE}; default) or whether all individual parameters must be specified (\code{=TRUE}).}

\item{msg}{A logical that indicates whether a message about the model and parameter definitions should be output (\code{=TRUE}) or not (\code{=FALSE}; default).}

\item{\dots}{Not implemented.}
}
\value{
\code{RichardsFuns} returns a function that can be used to predict fish length or weight given a vector of ages and values for the function parameters.  The result should be saved to an object that can then be used as a function name.  When the resulting function is used, the parameters are ordered as shown when the definitions of the parameters are printed after the function is called (if \code{msg=TRUE}).  If \code{simple=FALSE}, then the values for all parameters may be included as a vector in the first parameter argument.  If \code{simple=TRUE}, then all parameters must be declared individually.  The resulting function is somewhat easier to read when \code{simple=TRUE}.

\code{RichardsModels} returns a graphic that uses \code{\link{plotmath}} to show the model formulae in a pretty format.
}
\description{
Creates a function for a specific parameterization of the Richards growth function.  Use \code{RichardsModels()} to see the equations for each model.
}
\note{
Within FSA, Linf is the mean asymptotic length, ti is the age at the inflection point, k is related to growth (slope at the inflection point), and b is related to the position of the inflection point.

The parameterizations (1-4) corresponse to models/equations 1, 4, 5, and 6, respectively, in Tjorve and Tjorve (2010).  Note that their A, S, k, and d are Linf, a, k, and b here (in FSA).
}
\section{IFAR Chapter}{
 None specifically, but \href{https://fishr.wordpress.com/books/ifar/}{9-Individual Growth} is related.
}
\examples{
## See the formulae
\dontrun{windows(5,5)}
RichardsModels()

## Simple Examples
( rich1 <- RichardsFuns() )
ages <- 0:15
plot(rich1(ages,Linf=800,k=0.5,a=1,b=6)~ages,type="b",pch=19)

( rich2 <- RichardsFuns(2) )
plot(rich2(ages,Linf=800,k=0.5,ti=3,b=6)~ages,type="b",pch=19)

( rich3 <- RichardsFuns(3) )
plot(rich3(ages,Linf=800,k=0.5,ti=3,b=0.15)~ages,type="b",pch=19)

( rich4 <- RichardsFuns(4) )
plot(rich4(ages,Linf=800,k=0.5,ti=3,b=0.95)~ages,type="b",pch=19)
lines(rich4(ages,Linf=800,k=0.5,ti=3,b=1.5)~ages,type="b",pch=19,col="blue")

( rich2c <- RichardsFuns(2,simple=TRUE) ) # compare to rich2

#######################################################################################
## Examples of fitting Richards models
##   After the last example a plot is constructed with four lines on top of each
##   other illustrating that the parameterizations all produce the same fitted values.
# Make some fake data using the original parameterization

gompO <- gompFuns("original")
# setup ages, sample sizes (general reduction in numbers with
# increasing age), and additive SD to model
t <- 1:15
n <- c(10,40,35,25,12,10,10,8,6,5,3,3,3,2,2)
sd <- 15
# expand ages
ages <- rep(t,n)
# get lengths from gompertz and a random error for individuals
set.seed(333454)
lens <- gompO(ages,Linf=450,a=1,gi=0.3)+rnorm(length(ages),0,sd)
# put together as a data.frame
df <- data.frame(age=ages,len=round(lens,0))

# Fit first Richards parameterization
fit1 <- nls(len~rich1(age,Linf,k,a,b),data=df,start=list(Linf=450,k=0.25,a=0.65,b=3))
summary(fit1,correlation=TRUE)
plot(len~age,data=df,pch=19,col=rgb(0,0,0,1/5))
curve(rich1(x,Linf=coef(fit1)),from=0,to=15,col="red",lwd=10,add=TRUE)

# Fit second Richards parameterization
fit2 <- nls(len~rich2(age,Linf,k,ti,b),data=df,start=list(Linf=450,k=0.25,ti=3,b=3))
summary(fit2,correlation=TRUE)
curve(rich2(x,Linf=coef(fit2)),from=0,to=15,col="blue",lwd=7,add=TRUE)

# Fit third Richards parameterization
fit3 <- nls(len~rich3(age,Linf,k,ti,b),data=df,start=list(Linf=450,k=0.25,ti=3,b=-0.3))
summary(fit3,correlation=TRUE)
curve(rich3(x,Linf=coef(fit3)),from=0,to=15,col="green",lwd=4,add=TRUE)

# Fit fourth Richards parameterization
fit4 <- nls(len~rich4(age,Linf,k,ti,b),data=df,start=list(Linf=450,k=0.25,ti=3,b=0.7))
summary(fit4,correlation=TRUE)
curve(rich4(x,Linf=coef(fit4)),from=0,to=15,col="black",lwd=1,add=TRUE)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}, thanks to Gabor Grothendieck for a hint about using \code{get()}.
}
\references{
Richards, F. J.  1959.  A flexible growth function for empirical use.  Journal of Experimental Biology 10:290-300.

Tjorve, E. and K. M. C. Tjorve.  2010.  \href{https://www.researchgate.net/profile/Even_Tjorve/publication/46218377_A_unified_approach_to_the_Richards-model_family_for_use_in_growth_analyses_why_we_need_only_two_model_forms/links/54ba83b80cf29e0cb04bd24e.pdf}{A unified approach to the Richards-model family for use in growth analyses: Why we need only two model forms.}  Journal of Theoretical Biology 267:417-425.
}
\seealso{
See \code{\link{vbFuns}}, \code{\link{gompFuns}}, \code{\link{logisticFuns}}, and \code{\link{schnute}} for similar functionality for other models.
}
\keyword{manip}

