% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growthSchnute.R
\name{growthSchnute}
\alias{growthSchnute}
\alias{schnute}
\alias{schnuteModels}
\title{The four-parameter growth function from Schnute (1981).}
\usage{
schnute(t, case = 1, t1 = min(t), t3 = max(t), L1 = NULL, L3 = NULL,
  a = NULL, b = NULL)

schnuteModels(...)
}
\arguments{
\item{t}{A numeric vector of ages over which to model growth.}

\item{case}{A string that indicates the case of the Schnute growth function to use.}

\item{t1}{The (young) age that corresponds to \code{L1}.  Set to minimum value in \code{t} by default.}

\item{t3}{The (old) age that corresponds to \code{L3}.  Set to maximum value in \code{t} by default.}

\item{L1}{The mean size/length at \code{t1}.}

\item{L3}{The mean size/length at \code{t3}.}

\item{a}{A dimensionless parameter that is related to the time/age at the inflection point.}

\item{b}{A dimensionless parameter that is related to size/length at the inflection point.}

\item{\dots}{Not implemented.}
}
\value{
\code{schnute} returns a predicted size/length given the case of the function and the provided parameter values.

\code{schnuteModels} returns a graphic that uses \code{\link{plotmath}} to show the model formulae in a pretty format.
}
\description{
The four-parameter growth function from Schnute (1981).  Use \code{schnuteModels()} to see the equations for each model.
}
\section{IFAR Chapter}{
 None specifically, but \href{https://fishr.wordpress.com/books/ifar/}{9-Individual Growth} is related.
}
\examples{
## See the formulae
\dontrun{windows(5,5)}
schnuteModels()

## Simple examples
ages <- 1:15
s1 <- schnute(ages,case=1,L1=30,L3=400,a=0.3,b=1)
s2 <- schnute(ages,case=2,L1=30,L3=400,a=0.3,b=1)
s3 <- schnute(ages,case=3,L1=30,L3=400,a=0.3,b=1)
s4 <- schnute(ages,case=4,L1=30,L3=400,a=0.3,b=1)

plot(s1~ages,type="l",lwd=2)
lines(s2~ages,lwd=2,col="red")
lines(s3~ages,lwd=2,col="blue")
lines(s4~ages,lwd=2,col="green")
}
\author{
Derek H. Ogle.
}
\references{
Schnute, J.  1981.  A versatile growth model with statistical stable parameters.  Canadian Journal of Fisheris and Aquatic Sciences 38:1128-1140.
}
\seealso{
See \code{\link{vbFuns}}, \code{\link{gompFuns}}, and \code{\link{logisticFuns}} for similar functionality for other models.
}
\keyword{manip}

