% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hist.formula}
\alias{hist.formula}
\title{Creates separate histograms by levels.}
\usage{
\method{hist}{formula}(formula, data = NULL, main = "", right = FALSE,
  pre.main = "", xlab = names(DF)[1], ylab = "Frequency",
  same.breaks = TRUE, same.ylim = TRUE, ymax = NULL, col = "gray90",
  nrow = round(sqrt(num)), ncol = ceiling(sqrt(num)), byrow = TRUE, ...)
}
\arguments{
\item{formula}{A formula of class(formula).  See details.}

\item{data}{An optional data frame that contains the variables in the model.}

\item{main}{A character string to be used as the main title only when a single histogram is produced.}

\item{right}{A logical that indicates if the histogram bins are right-closed (left open) intervals (\code{=TRUE}) or not (\code{=FALSE}; default).}

\item{pre.main}{A character string to be used as a prefix for the main title.  See details.}

\item{xlab}{A character label for the x-axis.  Defaults to name of quantitative variable in the formula.}

\item{ylab}{A character label for the y-axis.  Defaults to \dQuote{Frequency}.}

\item{same.breaks}{A logical that indicates whether the same break positions should be used on each histogram.  Defaults to \code{TRUE}.}

\item{same.ylim}{A logicial that indicates whether the same limits for the y-axis should be used on each histogram.  Defaults to \code{TRUE}.  Ignored if \code{ylmts} is non-null.}

\item{ymax}{A single value that sets the maximum y-axis limit for each histogram or a vector of length equal to the number of groups that sets the maximum y-axis limit for each histogram separately.}

\item{col}{A string that indicates the color for the bars on the histogram.  Defaults to a light shade of gray (i.e., \code{"gray90"}).}

\item{nrow}{A numeric that contains the number of rows to use on the graphic.}

\item{ncol}{A numeric that contains the number of columns to use on the graphic.}

\item{byrow}{A logical that indicates if the histograms should fill rows first (\code{=TRUE} or columns first (\code{=FALSE}).}

\item{\dots}{Other arguments to pass through to the default \code{hist()}.}
}
\value{
Nothing is returned; however, a graphic is produced.
}
\description{
Creates separate histograms of a quantitative variable by levels of a factor variable.
}
\details{
The formula must be of the form \code{~quantitative}, \code{quantitative~1}, \code{quantitative~factor}, or \code{quantitative~factor*factor2} where \code{quantitative} is the quantitative variable to construct the histograms for and \code{factor} or \code{factor2} are factor variables that contain the levels for which separate histograms should be constructed.

If the formula is of the form \code{~quantitative} or \code{quantitative~1} then only a single histogram of the quantitative variable will be produced.  This allows \code{hist.formula()} to be used similarly to \code{hist()} but with a \code{data=} argument.

The function produces a single (but see below) graphic that consists of a grid on which the separate histograms are printed.  The rows and columns of this grid are determined to construct a plot that is as square as possible.  However, the rows and columns can be set by the user with the \code{nrow=} and \code{ncol=} arguments.  If the product of the number of rows and number of columns set by the user is less than the total number of histograms to be constructed then multiple pages of histograms will be produced (each requiring the user to click on the graph to go to the next graph).  The x-axis of each separate histogram will be labeled identically.  The default x-axis label is the name of the quantitative variable.  This can be changed by the user with the \code{xlab=} argument.


I changed the \code{right=} argument from that used in the base \code{hist()} so that right-open (left-closed) is the default.
}
\note{
Students often need to look at the distribution of a quantitative variable separated for different levels of a categorical variable.  One method for examining these distributions is with \code{boxplot(quantitative~factor)}.  Other methods use functions in \pkg{Lattice} and \pkg{ggplots2} but these packages have some learning \sQuote{overhead} for newbie students.  The formula notation, however, is a common way in R to tell R to separate a quantitative variable by the levels of a factor.  Thus, this function adds code for formulas to the generic \code{hist} function.  This allows newbie students to use a common notation (i.e., formula) to easily create multiple histograms of a quantitative variable separated by the levels of a factor.
}
\examples{
## Using the defaults
hist(Sepal.Length~Species,data=iris)

## Add x-labels and use a pre-fix on the main labels
hist(Sepal.Length~Species,data=iris,xlab="Sepal Length (cm)",pre.main="Species==")

## Use different breaks and different y-axis limits for each graph
hist(Sepal.Length~Species,data=iris,xlab="Sepal Length (cm)",same.breaks=FALSE,same.ylim=FALSE)

## Use same but user-controlled breaks for each graph
hist(Sepal.Length~Species,data=iris,xlab="Sepal Length (cm)",breaks=seq(4,8,1))

## Use same but user-controlled maximum value for y-axis
hist(Sepal.Length~Species,data=iris,xlab="Sepal Length (cm)",ymax=30)

## Control the organization of the 'grid' of histograms
hist(Sepal.Length~Species,data=iris,xlab="Sepal Length (cm)",nrow=1,ncol=3)

## Use right=FALSE & freq=FALSE arguments to demonstrate sending an argument used by base hist()
hist(Sepal.Length~Species,data=iris,xlab="Sepal Length (cm)",right=FALSE,freq=FALSE)

## Add a junk variable to the iris data set to show two factors on RHS
iris$junk <- factor(sample(c("A","B"),nrow(iris),replace=TRUE))
hist(Sepal.Length~Species*junk,data=iris,xlab="Sepal Length (cm)")

## Single histogram without grouping using formula notation
hist(~Sepal.Length,data=iris,xlab="Sepal Length (cm)",main="")
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}, but this implementation is largely a modification of the code provided by Marc Schwartz on the R-help mailing list on 1Jun07.
}
\seealso{
See base \code{hist} for related functionality and \code{multhist} in \pkg{plotrix} for similar functionality.
}
\keyword{hplot}

