% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ks2d.R
\name{ks2d1}
\alias{ks2d1}
\alias{plot.ks2d1}
\alias{print.ks2d1}
\title{Computes the D test statistic for the one-sample two-dimensional Kolmogrov-Smirnov test.}
\usage{
ks2d1(x1, y1, justD = FALSE, KSp = FALSE, divbylen = TRUE)

\method{print}{ks2d1}(x, ...)

\method{plot}{ks2d1}(x, xlab = NULL, ylab = NULL, xlim = range(x$x1),
  ylim = range(x$y1), pch = 16, ...)
}
\arguments{
\item{x1}{Vector of X values.}

\item{y1}{Vector of Y values.}

\item{justD}{Logical that indicates whether just the D test statistic (\code{=TRUE}) or more information should be returned (default; see value section below).}

\item{KSp}{Logical that indicates whether the approximate p-value from the Kolmogorov-Smirnov distribution should be returned.  HIGHLY EXPERIMENTAL (NOT YET TESTED).}

\item{divbylen}{Logical that indicates whether the proportion of values in each quadrant is computed by dividing by the sample size (\code{=TRUE}, default) or by the number of points that could be assigned to quadrants.}

\item{x}{An object returned from \code{d2ks1}.}

\item{xlab}{A string to label the x-axis.}

\item{ylab}{A string to label the y-axis.}

\item{xlim}{A vector of length two that indicates the limits over which to plot the x-axis.}

\item{ylim}{A vector of length two that indicates the limits over which to plot the y-axis.}

\item{pch}{A numeric that indicates the character to be used when plotting the results.}

\item{\dots}{Additional arguments sent to the plot function.}
}
\value{
The main function returns a single numeric of the D test statistic if \code{justD=TRUE} or a list with the following items if \code{justD=FALSE}:
 \itemize{
   \item D The D test statistic.  See details.
   \item pval The approximate p-value from the Kolmogorov-Smirnov distribution.  Returned only if \code{KSp=TRUE}.  HIGHLY EXPERIMENTAL (NOT YET TESTED).
   \item n1 Sample size.
   \item max1 Maximum D .
   \item where1 Observation(s) where maximum D occurred.
   \item KSp Same logical supplied by user.
   \item divbylen Same logical supplied by user.
   \item x1 Same vector of X data supplied by user.
   \item y1 Same vector of Y data supplied by user.
 }

The \code{plot} function return a scatterplot of the data with the point or points that resulted in the maximum D value.  The \code{print} function prints results in a nice format if \code{justD=FALSE} (only D will be printed if \code{justD=TRUE}).
}
\description{
Computes the D test statistic for the one-sample two-dimensional Kolmogrov-Smirnov test.
}
\details{
NEED DETAIL HERE.
}
\note{
This function is experimental at best at this point.
}
\examples{
# Example from Figure 4a in Garvey
data(Garvey4a)

# with the defaults
( res1 <- with(Garvey4a,ks2d1(shad,sunfish)) )
plot(res1,xlab="Shad Densitiy",ylab="Sunfish Density")

# same, but with modified computation of proportions
( res2 <- with(Garvey4a,ks2d1(shad,sunfish,divbylen=FALSE)) )

# same, but compute large-sample p-value -- HIGHLY EXPERIMENTAL (NOT TESTED)
( res3 <- with(Garvey4a,ks2d1(shad,sunfish,KSp=TRUE)) )

# Example from Figure 1 in Garvey
data(Garvey1)
( res4 <- with(Garvey1,ks2d1(Ameletus,Leuctra)) )
plot(res4,xlab="Ameletus Numbers",ylab="Leuctra Numbers",col=rgb(0,0,0,0.1))
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}, with significant help from Ben Bolker.
}
\references{
Garvey, J.E., E.A. Marschall, and R.A. Wright.  1998.  \href{http://opensiuc.lib.siu.edu/fiaq_pubs/17/}{From star charts to stoneflies: detecting relationships in continuous bivariate data.} Ecology 79:442 447.

Press, W.H., S.A. Teukolsky, W.T. Vetterling, B.P. Flannery.  2007.  \href{http://www.nr.com/}{Numerical Recipes: The Art of Scientific Computing, 3rd Edition.}  Cambridge University Press.  1286 pages.
}
\seealso{
\code{\link{ks2d1p}}
}
\keyword{htest}

