% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ks2d.R
\name{ks2d1p}
\alias{ks2d1p}
\alias{plot.ks2d1p}
\alias{print.ks2d1p}
\title{A permutation test to compute a p-value for the D test statistic for the one-sample two-dimensional Kolmogrov-Smirnov test.}
\usage{
ks2d1p(object, B = 100)

\method{print}{ks2d1p}(x, ...)

\method{plot}{ks2d1p}(x, xlab = "D Test Statistic", main = "", ...)
}
\arguments{
\item{object}{An object returned from \code{ks2d1}.}

\item{B}{A numeric representing the number of resamples.}

\item{x}{An object returned from \code{ks2d1p}.}

\item{xlab}{A string to label the x-axis.}

\item{main}{A string to label the main title on the plot.}

\item{\dots}{Additional arguments sent to the plot function.}
}
\value{
The main function returns a list with the following items:
 \itemize{
   \item D The D test statistic from the \code{ks2d2} object.
   \item pval The p-value from the permutation test.  See details.
   \item Ds The D test statistics from each of the B \sQuote{resamples}.
   \item B The \code{B} value supplied by the user.
 }
The \code{plot} function returns a density plot of the D test statistics from each of the B \sQuote{resamples} with the observed D test statistic shown with a vertical line.  The \code{print} function prints the results in a nice format).
}
\description{
A permutation test to compute a p-value for the D test statistic for the one-sample two-dimensional Kolmogrov-Smirnov test.
}
\details{
NEED DETAIL HERE.
}
\note{
This function is experimental at best at this point.
}
\examples{
data(Garvey4a)

# Results from the main 1-sample 2-D KS test from ks2d1
( res1 <- with(Garvey4a,ks2d1(shad,sunfish)) )
plot(res1,xlab="Shad Densitiy",ylab="Sunfish Density")

# permutation test using resampling
( res1p <- ks2d1p(res1,B=10) )  # B should be >1000, used 10 here to save time
plot(res1p)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}, with significant help from Ben Bolker.
}
\references{
Garvey, J.E., E.A. Marschall, and R.A. Wright.  1998.  \href{http://opensiuc.lib.siu.edu/fiaq_pubs/17/}{From star charts to stoneflies: detecting relationships in continuous bivariate data.} Ecology 79:442 447.

Press, W.H., S.A. Teukolsky, W.T. Vetterling, B.P. Flannery.  2007.  \href{http://www.nr.com/}{Numerical Recipes: The Art of Scientific Computing, 3rd Edition.}  Cambridge University Press.  1286 pages.
}
\seealso{
\code{\link{ks2d1}}
}
\keyword{htest}

