% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ks2d.R
\name{ks2d2}
\alias{ks2d2}
\alias{plot.ks2d2}
\alias{print.ks2d2}
\title{Computes the D test statistic for the two-sample two-dimensional Kolmogrov-Smirnov test.}
\usage{
ks2d2(x1, y1, x2, y2, justD = FALSE, KSp = FALSE, divbylen = TRUE)

\method{print}{ks2d2}(x, ...)

\method{plot}{ks2d2}(x, pchs = c(2, 19), cexs = c(1.25, 1), xlab = NULL,
  ylab = NULL, xlim = range(c(x$x1, x$x2)), ylim = range(c(x$y1, x$y2)),
  ...)
}
\arguments{
\item{x1}{Vector of X-coordinates for the first set of coordinates.}

\item{y1}{Vector of Y-coordinates for the first set of coordinates.}

\item{x2}{Vector of X-coordinates for the second set of coordinates.}

\item{y2}{Vector of Y-coordinates for the second set of coordinates.}

\item{justD}{Logical that indicates whether just the D test statistic (\code{=TRUE}) or more information should be returned (default; see value section below).}

\item{KSp}{Logical that indicates whether the approximate p-value from the Kolmogorov-Smirnov distribution should be returned.  HIGHLY EXPERIMENTAL (NOT YET TESTED).}

\item{divbylen}{Logical that indicates whether the proportion of values in each quadrant is computed by dividing by the sample size (\code{=TRUE}, default) or by the number of points that could be assigned to quadrants.}

\item{x}{An object returned from \code{d2ks}.}

\item{pchs}{A vector of length two plotting characters to be used when plotting the two sets of coordinates.}

\item{cexs}{A vector of length two character expansion values to be used when plotting the two sets of coordinates.}

\item{xlab}{A string to label the x-axis.}

\item{ylab}{A string to label the y-axis.}

\item{xlim}{A vector of length two that indicates the limits over which to plot the x-axis.}

\item{ylim}{A vector of length two that indicates the limits over which to plot the y-axis.}

\item{\dots}{Additiona arguments sent to the plot function.}
}
\value{
The main function returns a single numeric of the D test statistic if \code{justD=TRUE} or a list with the following items if \code{justD=FALSE}:
 \itemize{
   \item D The D test statistic.  See details.
   \item pval The approximate p-value from the Kolmogorov-Smirnov distribution.  Returned only if \code{KSp=TRUE}.  HIGHLY EXPERIMENTAL (NOT YET TESTED)
   \item n1 Sample size of first set of coordinates.
   \item n2 Sample size of second set of coordinates.
   \item max1 Maximum D for first set of coordinates.
   \item max2 Maximum D for second set of coordinates.
   \item where1 Observation(s) where maximum D occurred for first set of coordinates.
   \item where2 Observation(s) where maximum D occurred for second set of coordinates.
   \item KSp Same logical supplied by user.
   \item divbylen Same logical supplied by user.
   \item x1 Same vector of X-coordinates for the first set of coordinates supplied by user.
   \item y1 Same vector of Y-coordinates for the first set of coordinates supplied by user.
   \item x2 Same vector of X-coordinates for the second set of coordinates supplied by user.
   \item y2 Same vector of Y-coordinates for the second set of coordinates supplied by user.
 }

The \code{plot} function returns side-by-side plots showing the two sets of coordinates and the point or points that resulted in the maximum D value for each set of coordinates.  The \code{print} function prints results in a nice format if \code{justD=FALSE} (only D will be printed if \code{justD=TRUE}).
}
\description{
Computes the D test statistic for the two-sample two-dimensional Kolmogrov-Smirnov test.
}
\details{
NEED DETAIL HERE.
}
\note{
This function is experimental at best at this point.
}
\examples{
data(KS2D_NR)

# separate into the two sets of coordinates
d1 <- subset(KS2D_NR,group=="triangles")
d2 <- subset(KS2D_NR,group=="squares")

# perform analysis
( res1 <- ks2d2(d1$x,d1$y,d2$x,d2$y) )
plot(res1,xlim=c(-3,3),ylim=c(-3,3))
# same but with large sample p-values --  HIGHLY EXPERIMENTAL (NOT YET TESTED)
( res1 <- ks2d2(d1$x,d1$y,d2$x,d2$y,KSp=TRUE) )

# perform same analysis with modified computation of proportions
( res2 <- ks2d2(d1$x,d1$y,d2$x,d2$y,divbylen=FALSE) )
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}, with significant help from Ben Bolker.
}
\references{
Garvey, J.E., E.A. Marschall, and R.A. Wright.  1998.  \href{http://opensiuc.lib.siu.edu/fiaq_pubs/17/}{From star charts to stoneflies: detecting relationships in continuous bivariate data.} Ecology 79:442 447.

Press, W.H., S.A. Teukolsky, W.T. Vetterling, B.P. Flannery.  2007.  \href{http://www.nr.com/}{Numerical Recipes: The Art of Scientific Computing, 3rd Edition.}  Cambridge University Press.  1286 pages.
}
\seealso{
\code{\link{ks2d1}} and \code{\link{ks2d2p}}
}
\keyword{htest}

