% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logisticModels.R
\name{logisticModels}
\alias{logisticModels}
\title{Show the logistic growth function formulae implemented in the FSA package.}
\usage{
logisticModels(...)
}
\arguments{
\item{\dots}{Additional arguments for \code{\link{plot}}.  Generally not needed}
}
\value{
A graphic that uses \code{\link{plotmath}} to show the model formulae in a pretty format.
}
\description{
Show the logistic growth function formulae implemented in \code{\link{logisticFuns}}.
}
\section{IFAR Chapter}{
 None specifically, but \href{https://fishr.wordpress.com/books/ifar/}{9-Individual Growth} is related.
}
\examples{
\dontrun{windows(5,5)}
logisticModels()
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
See references in \code{\link{logisticFuns}}.
}
\seealso{
See \code{\link{logisticFuns}} for functions that represent the logistic growth function parameterizations.
}
\keyword{hplot}
\keyword{manip}

