% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{confint.nlsBoot}
\alias{confint.nlsBoot}
\alias{confint.nlsboot}
\alias{htest}
\alias{htest.nlsBoot}
\alias{predict.nlsBoot}
\title{Associated S3 methods for nlsBoot from nlstools.}
\usage{
\method{confint}{nlsBoot}(object, parm = NULL, level = 0.95,
  conf.level = level, plot = FALSE, err.col = "black", err.lwd = 2,
  rows = round(sqrt(ncol(object$coefboot))),
  cols = ceiling(sqrt(ncol(object$coefboot))), ...)

htest(object, ...)

\method{htest}{nlsBoot}(object, parm = NULL, bo = 0, alt = c("two.sided",
  "less", "greater"), plot = FALSE, ...)

\method{predict}{nlsBoot}(object, FUN, MARGIN = 1, conf.level = 0.95,
  digits = NULL, ...)
}
\arguments{
\item{object}{An object saved from \code{nlsBoot()}.}

\item{parm}{An integer that indicates which parameter to compute the confidence interval for.  Will compute for all parameters if \code{NULL}.}

\item{level}{Same as \code{conf.level}.  Used for compatability with the main \code{confint}.}

\item{conf.level}{A level of confidence as a proportion.}

\item{plot}{A logical that indicates whether a plot should be consturcted.  If \code{confint} then a histogram of the \code{parm} parameters from the bootstrap samples with error bars that illustrate the bootstrapped confidence intervals should be constructed.  If code{htest} then a histogram of the \code{parm} parameters with a vertical lines illustrating the \code{bo}value should be constructed.}

\item{err.col}{A single numeric or character that identifies the color for the error bar on the plot.}

\item{err.lwd}{A single numeric that identifies the line width for the error bar on the plot.}

\item{rows}{A numeric that contains the number of rows to use on the graphic.}

\item{cols}{A numeric that contains the number of columns to use on the graphic.}

\item{bo}{The null hypothesized parameter value.}

\item{alt}{A string that identifies the \dQuote{direction} of the alternative hypothesis.  See details.}

\item{FUN}{The function to be applied.}

\item{MARGIN}{A single numeric that indicates the margin over which to apply the function. \code{MARGIN=1} will apply to each row and is the default.}

\item{digits}{A single numeric that indicates the number of digits for the result.}

\item{\dots}{Additional arguments to functions.}
}
\value{
\code{confint} returns a matrix with as many rows as columns (i.e., parameter estimates) in the \code{object$coefboot} data frame and two columns of the quantiles that correspond to the approximate confidence interval.

\code{htest} returns a matrix with two columns.  The first column contains the hypothesized value sent to this function and the second column is the corresponding p-value.

\code{predict} returns a matrix with one row and three columns with the first column holding the predicted value (i.e., the median prediction) and the last two columns holding the approximate confidence interval.

\code{predict}
}
\description{
Provides S3 methods to construct non-parametric bootstrap confidence intervals and hypothesis tests for parameter values and predicted values of the response variable for a \code{\link[nlstools]{nlsBoot}} object (from the \pkg{nlstools} package).
}
\details{
\code{confint} finds the two quantiles that have the proportion (1-\code{conf.level})/2 of the bootstrapped parameter estimates below and above.  This is an approximate 100\code{conf.level}\% confidence interval.

In \code{htest} the \dQuote{direction} of the alternative hypothesis is identified by a string in the \code{alt=} argument.  The strings may be \code{"less"} for a \dQuote{less than} alternative, \code{"greater"} for a \dQuote{greater than} alternative, or \code{"two.sided"} for a \dQuote{not equals} alternative (the DEFAULT).  In the one-tailed alternatives the p-value is the proportion of bootstrapped parameter estimates in \code{object$coefboot} that are extreme of the null hypothesized parameter value in \code{bo}.  In the two-tailed alternative the p-value is twice the smallest of the proportion of bootstrapped parameter estimates above or below the null hypothesized parameter value in \code{bo}.

In \code{predict}, a user-supplied function is applied to each row of the \code{coefBoot} object in a \code{nlsBoot} object and then finds the median and the two quantiles that have the proportion (1-\code{conf.level})/2 of the bootstrapped predictions below and above.  The median is returned as the predicted value and the quantiles are returned as an approximate 100\code{conf.level}\% confidence interval for that prediction.
}
\examples{
data(Ecoli)
fnx <- function(days,B1,B2,B3) {
  if (length(B1) > 1) {
    B2 <- B1[2]
    B3 <- B1[3]
    B1 <- B1[1]
  }
  B1/(1+exp(B2+B3*days))
}
nl1 <- nls(cells~fnx(days,B1,B2,B3),data=Ecoli,start=list(B1=6,B2=7.2,B3=-1.45))
if (require(nlstools)) {
  nl1.boot <- nlsBoot(nl1,niter=99)  # way too few
  confint(nl1.boot,"B1")
  confint(nl1.boot,c(2,3))
  confint(nl1.boot,conf.level=0.90)
  htest(nl1.boot,1,bo=6,alt="less")
  predict(nl1.boot,fnx,days=3)
}
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
See \code{\link[nlstools]{summary.nlsBoot}} in \pkg{nlstools}
}
\keyword{htest}

