% Generated by roxygen2 (4.0.1): do not edit by hand
\name{odd}
\alias{odd}
\title{Determines if a number is odd.}
\usage{
odd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A logical vector of the same length as x.
}
\description{
Determines if a number is odd.
}
\examples{
d <- 1:8
odd(d)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{manip}

