% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oddeven.R
\name{is.odd}
\alias{is.even}
\alias{is.odd}
\title{Determines if a number is odd or even.}
\usage{
is.odd(x)

is.even(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A logical vector of the same length as x.
}
\description{
Determines if a number is odd or even.
}
\examples{
d <- 1:8
is.odd(d)
is.even(d)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{manip}

