\name{paletteChoices}
\alias{paletteChoices}
\title{Provides a vector of possible color palettes.}
\usage{
paletteChoices()
}
\value{
A single string with a palette name.
}
\description{
Provides a vector of possible color palettes.
}
\details{
Generally serves as an internal function and should be
called by the user.
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{misc}

