% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/perc.R
\name{perc}
\alias{perc}
\title{Computes the percentage of values in a vector less than or greater than (and equal to) some value.}
\usage{
perc(x, val, dir = c("greater", "less"), na.rm = TRUE,
  digits = getOption("digits"))
}
\arguments{
\item{x}{A numeric vector.}

\item{val}{A single numeric value.}

\item{dir}{A string that indicates whether the percentage is for values in \code{x} that are \code{"greater"} or \code{"less"} than \code{val}.}

\item{na.rm}{A logical that indicates whether \code{NA} values should be removed (DEFAULT) from \code{x} or not.}

\item{digits}{A single numeric that indicates the number of decimals the percentage should be rounded to.}
}
\value{
A single numeric that is the percentage of values in \code{x} that are \code{"greater"} or \code{"less"} than (and equal to) \code{val}.
}
\description{
Computes the percentage of values in a vector less than or greater than (and equal to) a user-supplied value.
}
\details{
This function is most useful when used with an apply-type of function.
}
\examples{
( tmp <- c(1:10,NA,NA) )
perc(tmp,5)
perc(tmp,5,"less")
perc(tmp,5,na.rm=FALSE)
perc(tmp,c(5,7),"less")
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{misc}

