% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RichardsStarts.R
\name{RichardsStarts}
\alias{RichardsStarts}
\title{Find reasonable starting values for a Richards growth function.}
\usage{
RichardsStarts(
  formula,
  data = NULL,
  param = 1,
  type = param,
  fixed = NULL,
  plot = FALSE,
  col.mdl = "gray70",
  lwd.mdl = 3,
  lty.mdl = 1,
  cex.main = 0.9,
  col.main = "red",
  ...
)
}
\arguments{
\item{formula}{A formula of the form \code{len~age}.}

\item{data}{A data frame that contains the variables in \code{formula}.}

\item{type, param}{A string that indicates the parameterization of the von Bertalanffy model.}

\item{fixed}{A named list that contains user-defined rather than automatically generated (i.e., fixed) starting values for one or more parameters. See details.}

\item{plot}{A logical that indicates whether a plot of the data with the superimposed model fit at the starting values should be created.}

\item{col.mdl}{A color for the model when \code{plot=TRUE}.}

\item{lwd.mdl}{A line width for the model when \code{plot=TRUE}.}

\item{lty.mdl}{A line type for the model when \code{plot=TRUE}.}

\item{cex.main}{A character expansion value for the main title when \code{plot=TRUE}.}

\item{col.main}{A color for the main title when \code{plot=TRUE}.}

\item{...}{Further arguments passed to the methods.}
}
\value{
A list that contains reasonable starting values. Note that the parameters will be listed in the same order and with the same names as listed in \code{\link{RichardsFuns}}.
}
\description{
Finds reasonable starting values for the parameters in a specific parameterization of the Richards growth function.
}
\details{
This function attempts to find reasonable starting values for a variety of parameterizations of the Richards growth function. There is no guarantee that these starting values are the \sQuote{best} starting values. One should use them with caution and should perform sensitivity analyses to determine the impact of different starting values on the final model results.

BLAH-BLAH-BLAH
}
\section{IFAR Chapter}{
 12-Individual Growth.
}

\examples{
# Make some fake data using the default parameterization
rfun <- RichardsFuns()
# setup ages, sample sizes (general reduction in numbers with
# increasing age), and additive SD to model
t <- 0:15
n <- c(5,10,40,35,25,12,10,10,8,6,5,3,3,3,2,2)
sd <- 15
# expand ages
ages <- rep(t,n)
# get lengths from gompertz and a random error for individuals
lens <- rfun(ages,Linf=450,ti=2,k=0.5,b1=1.5)+rnorm(length(ages),0,sd)
# put together as a data.frame
df <- data.frame(age=ages,tl=round(lens,0))

RichardsStarts(tl~age,data=df)
RichardsStarts(tl~age,data=df,plot=TRUE)

RichardsStarts(tl~age,data=df,param=1,plot=TRUE)
RichardsStarts(tl~age,data=df,param=2,plot=TRUE)
RichardsStarts(tl~age,data=df,param=3,plot=TRUE)
RichardsStarts(tl~age,data=df,param=4,plot=TRUE)
RichardsStarts(tl~age,data=df,param=5,plot=TRUE)

}
\seealso{
See \code{\link{growthFunShow}} to display the equations for the parameterizations used in \pkg{FSA} and \code{\link{RichardsFuns}} for functions that represent the logistic function parameterizations. See \code{\link{nlsTracePlot}} for help troubleshooting nonlinear models that don't converge.
}
\author{
Derek H. Ogle, \email{DerekOgle51@gmail.com}
}
\keyword{manip}
