% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/schnute.R
\name{schnute}
\alias{schnute}
\title{The four-parameter growth function from Schnute (1981).}
\usage{
schnute(t, case = 1, t1 = min(t), t3 = max(t), L1 = NULL, L3 = NULL,
  a = NULL, gamma = NULL)
}
\arguments{
\item{t}{A numeric vector of ages over which to model growth.}

\item{case}{A string that indicates the case of the Schnute growth function to use.}

\item{t1}{The (young) age that corresponds to \code{L1}.  Set to minimum value in \code{t} by default.}

\item{t3}{The (old) age that corresponds to \code{L3}.  Set to maximum value in \code{t} by default.}

\item{L1}{The mean size/length at \code{t1}.}

\item{L3}{The mean size/length at \code{t3}.}

\item{a}{A dimensionless parameter that is related to the time/age at the inflection point.}

\item{gamma}{A dimensionless parameter that is related to size/length at the inflection point.}
}
\value{
A predicted size/length given the case of the function and the provided parameter values..
}
\description{
The four-parameter growth function from Schnute (1981).
}
\section{IFAR Chapter}{
 None specifically, but \href{https://fishr.wordpress.com/books/ifar/}{9-Individual Growth} is related.
}
\examples{
ages <- 1:15
s1 <- schnute(ages,case=1,L1=30,L3=400,a=0.3,gamma=1)
s2 <- schnute(ages,case=2,L1=30,L3=400,a=0.3,gamma=1)
s3 <- schnute(ages,case=3,L1=30,L3=400,a=0.3,gamma=1)
s4 <- schnute(ages,case=4,L1=30,L3=400,a=0.3,gamma=1)

plot(s1~ages,type="l",lwd=2)
lines(s2~ages,lwd=2,col="red")
lines(s3~ages,lwd=2,col="blue")
lines(s4~ages,lwd=2,col="green")
}
\author{
Derek H. Ogle.
}
\references{
Schnute, J.  1981.  A versatile growth model with statistical stable parameters.  Canadian Journal of Fisheris and Aquatic Sciences 38:1128-1140.
}
\seealso{
See \code{\link{schnuteModels}} for equations for each case.
}
\keyword{manip}

