% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/schnuteModels.R
\name{schnuteModels}
\alias{schnuteModels}
\title{Show the cases for the Schnute growth function formulae implemented in the FSA package.}
\usage{
schnuteModels(...)
}
\arguments{
\item{\dots}{Additional arguments for \code{\link{plot}}.  Generally not needed}
}
\value{
A graphic that uses \code{\link{plotmath}} to show the model formulae in a pretty format.
}
\description{
Show the cases for the Schnute growth function formulae implemented in \code{\link{schnuteFuns}}.
}
\section{IFAR Chapter}{
 None specifically, but \href{https://fishr.wordpress.com/books/ifar/}{9-Individual Growth} is related.
}
\examples{
\dontrun{windows(5,5)}
schnuteModels()
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
Schnute, J.  1981.  A versatile growth model with statistical stable parameters.  Canadian Journal of Fisheris and Aquatic Sciences 38:1128-1140.
}
\seealso{
See \code{\link{schnute}} for functions that represent the Schnute growth function cases.
}
\keyword{hplot}
\keyword{manip}

