% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGrowthFun.R
\name{showGrowthFun}
\alias{showGrowthFun}
\title{Creates an expression for a specific parameterization of the von Bertalanffy, Gompertz, Richards, and logistic growth functions, as well as the Schnute and Schnute-Richards growth functions.}
\usage{
showGrowthFun(
  type = c("von Bertalanffy", "Gompertz", "Richards", "logistic", "Schnute",
    "Schnute-Richards"),
  param = 1,
  pname = NULL,
  case = NULL,
  plot = FALSE,
  ...
)
}
\arguments{
\item{type}{A string (\dQuote{von Bertalanffy}, \dQuote{Gompertz}, \dQuote{logistic}, \dQuote{Richards}, \dQuote{Schnute}, \dQuote{Schnute-Richards}) that indicates the type of growth function to show.}

\item{param}{A single numeric that indicates the specific parameterization of the growth function. See details.}

\item{pname}{A single character that indicates the specific parameterization of the growth function. If \code{NULL} then \code{param} will be used. See details.}

\item{case}{A numeric that indicates the specific case of the Schnute function to use. See details.}

\item{plot}{A logical for whether the expression should be shown on a \dQuote{blank} plot. See examples.}

\item{\dots}{Arguments for \code{plot}. In particular use \code{cex=} to make the expression larger and easier to read. See examples.}
}
\value{
An expression representing the equation of the growth function given in \code{type} and \code{param}.
}
\description{
Creates an expression for a specific parameterization of the von Bertalanffy, Gompertz, Richards, and logistic growth functions, as well as the Schnute and Schnute-Richards growth functions. The expression can be plotted with \code{plot=TRUE} to see the equation of the growth function.
}
\examples{
#===== The expression 
showGrowthFun()
showGrowthFun(pname="Typical")

#===== Shown on the plot, and then larger
showGrowthFun(plot=TRUE)
showGrowthFun(plot=TRUE,cex=2)

#===== Examples of other growth odels
showGrowthFun(type="Richards",param=3,plot=TRUE,cex=1.5)
showGrowthFun(type="Schnute",case=2,plot=TRUE,cex=1.5)

#===== Put multiple expressions in one plot
op <- par(mar=c(0.1,0.1,0.1,0.1))
plot(0,type="n",xlab="",ylab="",xlim=c(0,1),ylim=c(0,3),xaxt="n",yaxt="n")
text(0,2.5,"Original:",pos=4)
text(0.5,2.5,showGrowthFun(type="von Bertalanffy",pname="Original"))
text(0,1.5,"Typical:",pos=4)
text(0.5,1.5,showGrowthFun(type="von Bertalanffy",pname="Typical"))
text(0,0.5,"Francis:",pos=4)
text(0.5,0.5,showGrowthFun(type="von Bertalanffy",pname="Francis"))
par(op)

#===== Put expression in title or otherwise on the plot
# Make a von Bertalanffy function
vb1 <- makeGrowthFun()
# make vector of ages over which to plot
ages <- 1:20
# Get and save the expression of the von Bertalanffy growth function
tmp <- showGrowthFun()

# Use expression as title on plot
plot(vb1(ages,Linf=20,K=0.3,t0=-0.2)~ages,type="b",pch=19,main=tmp,ylab="Length")

# Put expression in the plot body
text(15,10,tmp)

}
\seealso{
See \code{\link{makeGrowthFun}} to make functions that correspond to these expressions.
}
