% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{SpotVA1}
\alias{SpotVA1}
\title{Age and length of spot.}
\format{A data frame of 403 observations on the following 2 variables:
\describe{
\item{tl}{Measured total lengths (in inches).}
\item{age}{Ages assigned from examination of otoliths.}
}}
\source{
Extracted from Table 1 in Chapter 8 of the \href{http://ww2.odu.edu/sci/cqfe/}{Virginia Marine Resources Commissions's} Final Report on finfish ageing, 2002.
}
\description{
Ages (from otoliths) and length of Virginia Spot (\emph{Leiostomus xanthurus}).
}
\details{
Final length measurements were simulated by adding a uniform error to the value at the beginning of the length category.
}
\section{Topic(s)}{
 \itemize{
\item Growth
\item von Bertalanffy
}
}
\examples{
data(SpotVA1)
str(SpotVA1)
head(SpotVA1)
plot(tl~age,data=SpotVA1)
}
\seealso{
\code{SpotVA2} in \pkg{FSAdata}.
}
\concept{
Growth 'von Bertalanffy'
}
\keyword{datasets}

