% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/srModels.R
\name{srModels}
\alias{srModels}
\title{Show the stock-recruitment model formulas implemented in FSA.}
\usage{
srModels(...)
}
\arguments{
\item{\dots}{Additional arguments for \code{plot}.  Generally not needed.}
}
\value{
A graphic that uses \code{plotmath} to show the model formulas in a pretty format.
}
\description{
Show the stock-recruitment model formulas implemented in \code{\link{srFuns}}, \code{\link{srStarts}}, and \code{srSim}.
}
\section{IFAR Chapter}{
 \href{https://fishr.wordpress.com/books/ifar/}{10-Recruitment}.
}
\examples{
\dontrun{windows(6,5)}
srModels()
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
Ogle, D.H.  2016.  Introductory Fisheries Analyses with R.  Chapman & Hall/CRC, Boca Raton, FL.
}
\seealso{
See \code{\link{srFuns}} and \code{\link{srStarts}} for related functionality.
}
\keyword{hplot}
\keyword{manip}

