% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Summarize.R
\name{Summarize}
\alias{Summarize}
\alias{Summarize.default}
\alias{Summarize.formula}
\title{Summary statistics for a numeric or factor variable.}
\usage{
Summarize(object, ...)

\method{Summarize}{default}(object, digits = getOption("digits"),
  addtotal = TRUE, percent = TRUE, percdigs = 2, na.rm = TRUE,
  exclude = "", ...)

\method{Summarize}{formula}(object, data = NULL,
  digits = getOption("digits"), percent = c("row", "column", "total",
  "none"), percdigs = 2, addtotal = TRUE, na.rm = TRUE, exclude = "",
  ...)
}
\arguments{
\item{object}{A vector of numeric or factor data.}

\item{digits}{A numeric that indicates the number of decimals to round the numeric summaries to.}

\item{addtotal}{A logical that indicates whether totals should be added to tables (\code{=TRUE}, default) or not.}

\item{percent}{A logical that indicates whether frequency tables should include percentages (\code{=TRUE}, default) or not for a single categorical variable or a string that indicates the type of percents to compute for a two-way table constructed wtih the formula.}

\item{percdigs}{A numeric that indicates the number of decimals to round the percentage summaries to.}

\item{na.rm}{A logical that indicates whether numeric missing values (\code{NA}) should be removed (\code{=TRUE}, default) or not.}

\item{exclude}{A string that contains the code that should be excluded from the levels of the factor variable.}

\item{data}{An optional data frame that contains the variables in the model.}

\item{\dots}{Other arguments to the generic \code{summary}, \code{sd}, or \code{table} functions.}
}
\value{
A named vector or data frame (when a quantitative variable is separted by one or two factor variables) of summary statistics for numeric data and a matrix of frequencies and, possibly, percentages for factor variables.
}
\description{
Summary statistics for a single numeric or factor variable, possibly separated by the levels of a factor variable.  Very similar to \code{summary} for a numeric variables and \code{table} for factor variables.
}
\details{
For numeric data this is the same as \code{summary} except that \code{Summarize} includes the sample size, valid sample size (sample size minus number of \code{NA}s) and standard deviation (i.e., \code{sd}).  Also the output is ordered slightly differently.

For a factor variable this function computes a frequency table, a percentage table (if \code{percent=TRUE}), and a valid percentage table (percentage if \dQuote{NA}s are excluded; if \code{percent=TRUE}).  The tables will contain a total row if \code{addtotal=TRUE}.

The \code{object} argument can be a formula of form \code{y~x} where \code{y} can be either a numeric or factor variable and \code{x} can be only a factor variable or \code{y~x*z} where \code{z} is a second factor variable.  More complicated formulas are not supported.  When \code{y} is numeric then the summary statistics of \code{y} will be computed for each level in \code{x} or the combinations of the levels of \code{x} and \code{z}.  When \code{y} is a factor then a two-way table will be computed.  If \code{addtotal=TRUE} then row totals only will be added.  If \code{percent=TRUE} then a row percentages table will be computed such that the percentages represent the percent in the levels of \code{x} for each level of \code{y}.
}
\note{
The \code{Summarize} function, when applied to a vector of quantitative data, produces results of basic statistics similar to that provided by \code{summary}.  The primary addition in the results of \code{Summarize} is the inclusion of the standard deviation and the (potential) calculation of a valid sample size.  When applied to a vector of categorical data, \code{Summarize} produces a frequency table with (by default) percentages and (perhaps) valid percentages.  The results for categorical data are NOT meant to replace the \code{table} function but to provide an alternative and to provide a useful result if the student provides it with categorical data.

Students often need to look at basic statistics of a quantitative variable separated for different levels of a categorical variable.  This type of analysis can be made with \code{tapply}, \code{by}, or \code{aggregate} (or a few other functions in other packages) but the use of these functions are not obvious to newbie students or return results in a format that is not obvious to newbie students.  Thus, the formula method to the \code{Summarize} generic function allows newbie students to use a common notation (i.e., formula) to easily compute summary statistics for a quantitative variable separated by the levels of a factor.
}
\examples{
## Create a numeric vector (with missing values)
n <- 102
y <- c(0,0,NA,NA,NA,runif(n-5))
## Create a factor vector (with missing values)
g1 <- factor(sample(c("A","B","C","NA"),n,replace=TRUE))
## Create a factor vector with unknowns
g2 <- factor(sample(c("male","female","UNKNOWN"),n,replace=TRUE))
# Put into a data.frame (with some extra variables)
d <- data.frame(dy=y,dg1=g1,dg2=g2,
                dw=sample(1:3,n,replace=TRUE),
                dv=sample(1:3,n,replace=TRUE))

# typical output of summary() for a numeric variable
summary(y)

# this function
Summarize(y)
Summarize(~dy,data=d)
Summarize(dy~1,data=d)
# control the number of digits
Summarize(~dy,data=d,digits=3)

## Factor vector (excluding "NA"s in second call)
Summarize(~dg1,data=d)
Summarize(~dg1,data=d,exclude="NA")

## Factor vector with UNKNOWNs
Summarize(~dg2,data=d)
Summarize(~dg2,data=d,exclude="UNKNOWN")

## Numeric vector by levels of a factor variable
Summarize(dy~dg1,data=d,digits=3)
Summarize(dy~dg1,data=d,digits=3,exclude="NA")
Summarize(dy~dg2,data=d,digits=3)
Summarize(dy~dg2,data=d,digits=3,exclude="UNKNOWN")

## What happens if RHS of formula is not a factor
Summarize(dy~dw,data=d)
Summarize(y~dw*dv,data=d)

## Summarize factor variable by a factor variable
Summarize(dg1~dg2,data=d)
Summarize(dg1~dg2,data=d,exclude="NA")
Summarize(dg1~dg2,data=d,exclude=c("NA","UNKNOWN"))
Summarize(dg1~dg2,data=d,percent="none")
Summarize(dg1~dg2,data=d,percent="column")
Summarize(dg1~dg2,data=d,percent="total")

## Summarizing all variables in a data frame
lapply(as.list(d),Summarize,digits=4)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
See \code{\link{summary}}, \code{\link{table}}, and \code{\link{xtabs}} for related one dimenstional functionality.  See \code{\link{tapply}}, \code{\link[doBy]{summaryBy}} in \pkg{doBy}, \code{\link[psych]{describe}} in \pkg{psych}, \code{\link[prettyR]{describe}} in \pkg{prettyR}, and \code{\link[fBasics]{basicStats}} in \pkg{fBasics} for similar \dQuote{by} functionality.
}
\keyword{misc}

