% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tictactoeAdd}
\alias{tictactoeAdd}
\title{Adds point and confidence intervals for observed preator-prey PSD to
tic-tac-toe graph.}
\usage{
tictactoeAdd(predval, preyval, conf.level = 0.95, ci.type = c("wilson",
  "exact", "asymptotic"), ci.lwd = 2, pt = 19, pt.col = "red",
  pt.cex = 1, label = NULL, lbl.pos = c("NE", "NW", "SE", "SW"),
  lbl.col = pt.col, lbl.cex = 1)
}
\arguments{
\item{predval}{A vector of length 2 that contains the number of predators in
two length categories.  It is assumed that the number in the larger length
category and, thus, the smaller number of individuals is given first.}

\item{preyval}{A vector of length 2 that contains the number of prey in two
length categories.  It is assumed that the number in the larger length
category and, thus, the smaller number of individuals is given first.}

\item{conf.level}{A numeric representing the confidence level for the confidence intervals.}

\item{ci.type}{A string representing the method to use for computing the
binomial confidence intervals.  See details of \code{binCI}.}

\item{ci.lwd}{A numeric designating the line width for the condince interval lines.}

\item{pt}{A numeric representing the type of symbol to use for the
predator/prey PSD point.}

\item{pt.col}{A string representing the color of the symbol used for the
predator/prey PSD point.}

\item{pt.cex}{A numeric representing the relative size of the symbol used for
the predator/prey PSD point.}

\item{label}{A string used to label the point.}

\item{lbl.pos}{A string that indicates where to place the label.}

\item{lbl.col}{A string that indicates the color to use for the label.  Defaults
to same color as the color of the point.}

\item{lbl.cex}{A numberic representing the relative size of the point label.}
}
\value{
None.  However, a base graphic is modified.
}
\description{
Adds point and confidence intervals for observed preator-prey PSD to
tic-tac-toe graph.  The base tic-tac-toe graph is constructed with the
\code{tictactoe} function.
}
\details{
This function adds the point and confidence interval for observed
predator-prey PSD values to a base tic-tac-toe graph that already exists.
Thus, the \code{tictactoe} function must be run before this function and the
resultant plot must be active.

Confidence intervals are constructed assuming a binomial distribution with
the \code{binCI} function.
}
\section{fishR vignette}{
 \url{https://sites.google.com/site/fishrfiles/gnrl/SizeStructure.pdf}
}
\examples{
## See example for tictactoe() function.
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
\code{\link{tictactoe}}, \code{\link{psdVal}}, \code{\link{binCI}}.
}
\keyword{aplot}

