% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validn}
\alias{validn}
\title{Finds the number of valid (non-NA) values in a vector.}
\usage{
validn(object)
}
\arguments{
\item{object}{A vector.}
}
\value{
A single numeric value that is the number of non-\code{NA} values in a vector.
}
\description{
Finds the number of valid (non-NA) values in a vector.
}
\examples{
junk1 <- c(1,7,2,4,3,10,NA)
junk2 <- c("Derek","Hugh","Ogle","Santa","Claus","Frank",NA,NA)
junk3 <- factor(junk2)
junk4 <- c(TRUE,TRUE,FALSE,FALSE,FALSE,TRUE,NA,NA)

validn(junk1)
validn(junk2)
validn(junk3)
validn(junk4)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
See \code{valid.n} in \pkg{plotrix} and \code{nobs} in \pkg{gdata}.  See \code{is.na} for finding the missing values.
}
\keyword{manip}

