% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vbModels.R
\name{vbModels}
\alias{vbModels}
\title{Show the von Vertalanffy model formulas implemented in the FSA package.}
\usage{
vbModels(...)
}
\arguments{
\item{\dots}{Additional arguments for \code{plot}.  Generally not needed}
}
\value{
A graphic that uses \code{plotmath} to show the model formulas in a pretty format.
}
\description{
Show the von Bertalanffy model formulas implemented in \code{\link{vbFuns}} and \code{\link{vbStarts}}.
}
\examples{
\dontrun{windows(8,5)}
vbModels()
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
See \code{\link{vbFuns}} for functions that represent the von Bertalanffy parameterizations and \code{\link{vbStarts}} for methods to find starting values.
}
\keyword{hplot}
\keyword{manip}

