% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WhitefishLC}
\alias{WhitefishLC}
\title{Assessed ages from two readers on three structures for Lake Whitefish from Lake Champlain.}
\format{A data frame with 151 observations on the following 11 variables:
 \describe{
   \item{fishID}{A unique fish identification number.}
   \item{tl}{Total length (in mm).}
   \item{scale1}{Assessed age from scales by first reader.}
   \item{scale2}{Assessed age from scales by second reader.}
   \item{scaleC}{Consensus age from scales by both reader.}
   \item{finray1}{Assessed age from fin rays by first reader.}
   \item{finray2}{Assessed age from fin rays by second reader.}
   \item{finrayC}{Consensus age from fin rays by both reader.}
   \item{otolith1}{Assessed age from otoliths by first reader.}
   \item{otolith2}{Assessed age from otoliths by second reader.}
   \item{otolithC}{Consensus age from otoliths by both reader.}
 }}
\source{
Data from Herbst, S.J. and J.E. Marsden.  2011.  \href{http://www.uvm.edu/rsenr/emarsden/Herbst\%20and\%20Marsden\%20whitefish\%20age\%20structure\%20comparison.pdf}{Comparison of precision and bias of scale, fin ray, and otolith age estimates for lake whitefish (\emph{Coregonus clupeaformis}) in Lake Champlain.}  Journal of Great Lakes Research.  37:386-389.  Contributed by Seth Herbst.  \bold{Do not use for other than educational purposes without permission from the author.}
}
\description{
Assessed ages from two readers on three structures for Lake Whitefish (\emph{Coregonus clupeaformis}) from Lake Champlain in 2009.
}
\section{Topic(s)}{

 \itemize{
   \item Age
   \item Ageing
   \item Precision
   \item Bias
 }
}
\examples{
data(WhitefishLC)
str(WhitefishLC)
head(WhitefishLC)
}
\concept{
Age Precision Bias Error
}
\keyword{datasets}

