% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsVal.R
\name{wsVal}
\alias{wsVal}
\title{Finds standard weight equation coefficients for a particular species.}
\usage{
wsVal(
  species = "List",
  thesaurus = NULL,
  group = NULL,
  units = c("metric", "English"),
  ref = NULL,
  method = NULL,
  simplify = FALSE,
  dat = NULL
)
}
\arguments{
\item{species}{A string that contains the species name for which to find Ws coefficients. See details.}

\item{thesaurus}{A named vector or list for providing alternative species names (the values in the list) that correspond to specific names in \code{WSlit} (the names in the list). See details and examples.}

\item{group}{A string that contains the sub-group of \code{species} for which to find the Ws coefficients; e.g., things like \dQuote{lotic}, \dQuote{lentic}, \dQuote{female}, \dQuote{male}.}

\item{units}{A string that indicates whether the coefficients for the standard weight equation to be returned are in \code{"metric"} (DEFAULT; mm and g) or \code{"English"} (in and lbs) units.}

\item{ref}{A numeric that indicates which percentile the equation should be returned for. Note that the vast majority of equations only exist for the \code{75}th percentile (DEFAULT).}

\item{method}{A string that indicates which equation-derivation method should be used (one of \code{"RLP"}, \code{"EmP"}, or \code{"Other"}). Defaults to \code{NULL} which will result in the only method available being returned or an error asking the user to choose a method for equations for which more than one method is available (which is the case for very few species).}

\item{simplify}{A logical that indicates whether the \sQuote{units}, \sQuote{ref}, \sQuote{measure}, \sQuote{method}, \sQuote{comments}, and \sQuote{source} fields should be included (\code{=FALSE}) or not (\code{=TRUE}; DEFAULT). See details.}

\item{dat}{Data.frame of Gabelhouse length categories for all species. Defaults to `WSlit` and is generally not used by the user (this simplifies use of this function in \code{psdAdd}).}
}
\value{
A one row data frame from \code{\link{WSlit}} that contains all known information about the standard weight equation for a given species, type of measurement units, and reference percentile if \code{simplify=FALSE}. If \code{simplify=TRUE} then only the species; minimum and maximum length for which the standard equation should be applied; and intercept, slope, and quadratic  coefficients for the standard weight equation. Note that the maximum length and the quadratic coefficient will not be returned if they do not exist in \code{\link{WSlit}} for the species.

If no arguments are given to this function then a list of available species names in \code{\link{WSlit}} will be printed. If the species name is mis-spelled (or mis-capitalized), multiple standard weight equations exist for the species (such that \code{group}, \code{ref}, or \code{method} should be used), or if a standard weight equation does not exist for the species in \code{\link{WSlit}}, then an error will be issued.
}
\description{
Returns a vector that contains information about the standard weight equation for a given species, including type of measurement units, reference percentile, method used to derive the equation, and literature source.
}
\details{
This function extracts all known information from \code{\link{WSlit}} about the following standard weight equation,

\deqn{log_{10}(Ws) = log_{10}(a) + blog_{10}(L) + blog_{10}(L)^{2}}

See \code{\link{WSlit}} for more information about the meaning of each value returned.

Note from above that the coefficients are returned for the TRANSFORMED model. Thus, to obtain the standard weight (Ws), the returned coefficients are used to compute the common log of Ws which must then be raised to the power of 10 to compute the Ws.

The \code{thesaurus} argument may be used to relate alternate species names to the species names used in \code{WSlit}. For example, you (or your data) may use \dQuote{Bluegill Sunfish}, but \dQuote{Bluegill} is used in \code{WSlit}. The alternate species name can be used here if it is defined in a named vector (or list) given to \code{thesarus=}. The alternate species name is the value and the species name in \code{PSDlit} is the name in this vector/list - e.g., \code{c("Bluegill"="Bluegill Sunfish")}. See the examples for a demonstration.

Some species have length categories separated by sub-group. For example, length categories exist for both lentic and lotic populations of Brown Trout. The length values for a sub-group may be obtained by either including the species name in \code{species} and the sub-group name in \code{group} or by using the combined species and sub-group name, with the sub-group name in parentheses, in \code{species}. Both methods are demonstrated in the examples. Note that an error is returned if a species has sub-groups but neither method is used to define the sub-group.

See examples and \href{https://fishr-core-team.github.io/FSA/articles/Computing_Relative_Weights.html}{this article} for a demonstration.
}
\section{IFAR Chapter}{
 8-Condition.
}

\examples{
#===== List all available Ws equations
wsVal()

#===== Find equations for Yellow Perch, in different formats
wsVal("Yellow Perch")
wsVal("Yellow Perch",units="metric")   # same as default
wsVal("Yellow Perch",units="English")
wsVal("Yellow Perch",units="English",simplify=TRUE)

#===== Find equation for Ruffe, demonstrating quadratic formula
wsVal("Ruffe",units="metric",ref=75,simplify=TRUE)
wsVal("Ruffe",units="metric",ref=50,simplify=TRUE)

#===== Find equation for Brown Trout, which has equations for sub-groups
#-----   demonstrating use of group= argument
wsVal("Brown Trout",group="lotic")
wsVal("Brown Trout",group="lentic")
#-----   demonstrating group combined in species name, so no group= arg
wsVal("Brown Trout (lotic)")
wsVal("Brown Trout (lentic)")

#===== Find equation for Bluegill Sunfish (aka Bluegill in WSlit)
#        Note that this not that useful here as could just use "Bluegill"
#        in first argument. This will be useful in wrAdd()
wsVal("Bluegill Sunfish",thesaurus=c("Bluegill"="Bluegill Sunfish"))

#===== Add Ws & Wr values to a data frame (for one species) ... also see wrAdd()
#----- Example data from PSDWRtest, simplify variables for this example
yepdf <- subset(PSDWRtest,species=="Yellow Perch",select=c("species","len","wt"))
str(yepdf)

#----- Get Ws equation info
( wsYEP <- wsVal("Yellow Perch",units="metric") )

#----- Add Ws (eqn is on log10-log10 scale ... so log10 length, 10^ result)
yepdf$ws <- 10^(wsYEP[["int"]]+wsYEP[["slope"]]*log10(yepdf$len))

#----- Change Ws for fish less than min.TL to NA
yepdf$ws[yepdf$len<wsYEP[["min.TL"]]] <- NA

#----- Add Wr
yepdf$wr <- yepdf$wt/yepdf$ws*100

#----- Examine results
peek(yepdf,n=6)

#----- Same as above but using dplyr
if (require(dplyr)) {
  yepdf <- PSDWRtest |> filter(species=="Yellow Perch") |> select(species,len,wt) |>
    mutate(ws=10^(wsYEP[["int"]]+wsYEP[["slope"]]*log10(len)),
           ws=ifelse(len<wsYEP[["min.TL"]],NA,ws),
           wr=wt/ws*100)
  peek(yepdf,n=6)
}

}
\references{
Ogle, D.H. 2016. \href{https://fishr-core-team.github.io/fishR/pages/books.html#introductory-fisheries-analyses-with-r}{Introductory Fisheries Analyses with R}. Chapman & Hall/CRC, Boca Raton, FL.
}
\seealso{
See \code{\link{wrAdd}} and \code{\link{WSlit}} for related functionality.
}
\author{
Derek H. Ogle, \email{DerekOgle51@gmail.com}
}
\keyword{manip}
