% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ageKeyPlot}
\alias{ageKeyPlot}
\title{Plots to visualize age-length keys.}
\usage{
ageKeyPlot(key, type = c("barplot", "area", "lines", "splines", "bubble"),
  xlab = "Length", ylab = "Proportion", showLegend = FALSE,
  lbl.cex = 1.25, leg.cex = 1, lwd = 2, span = 0.25, grid = TRUE,
  col = "gray80", buf = 0.45, bubble.ylab = "Age", add = FALSE,
  xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{key}{A numeric matrix that contains the age-length key.}

\item{type}{A string that indicates the type of plot to construct.  See details.}

\item{xlab}{A string that contains the label for the x-axis.}

\item{ylab}{A string that contains the label for the y-axis in all plots except when \code{type="bubble"} (see \code{bubble.ylab=)}.}

\item{bubble.ylab}{A string that contains the label for the y-axis when \code{type="bubble"}.}

\item{xlim}{A numeric of length 2 that provide the limits for the x-axis.}

\item{ylim}{A numeric of length 2 that provide the limits for the y-axis.}

\item{showLegend}{A logical that indicates whether a legend should be displayed for when \code{type="barplot"} and \code{type="area"}.  See examples.}

\item{lbl.cex}{A numeric character expansion value for labels inside the bars when \code{type="barplot"} and \code{showLegend=FALSE} or on the lines when \code{type="lines"} or \code{type="splines"}.}

\item{leg.cex}{A numeric character expansion value for labels on the legend when \code{type="barplot"} or \code{type="area"} and \code{showLegend=TRUE}.}

\item{lwd}{A numeric that indicates the line width for when \code{type="lines"} or \code{type="splines"}.}

\item{span}{A numeric that indicates the span value to use in \code{loess} when \code{type="splines"}.}

\item{grid}{A logical that indicates whether a grid should be placed under the bubbles when \code{type="bubble"} or a character or appropriate vector that identifies a color for the grid.  See examples.}

\item{col}{A string that indicates the color of the bubbles when \code{type="bubble"}.}

\item{buf}{A single numeric that indicates the relative width of the bubbles when \code{type="bubble"}. A value of 0.5 would mean that two full-width bubbles would touch each other either in the x- or y-direction (i.e., this would represent half of the minimum of the physical distance between values one-unit apart on the x- and y-axes).  Set this to a value less than 0.5 so that the bubbles will not touch (the default is 0.45).}

\item{add}{A logical that indicates whether the data should be added to an already existing plot.  May be useful for visually comparing age-length keys.  Only implemented when \code{type="bubble"}.}

\item{\dots}{Additional arguments to pass to \code{plot}, \code{barplot}, or \code{stackpoly}.}
}
\value{
None, but a plot is constructed.
}
\description{
Various plots to visualize the proportion of fish of certain ages within length categories in an age-length key.
}
\details{
A variety of plots can be used to visualize the proportion of fish of certain ages withing length categories of an age-length key.  The types of plots are described below and illustrated in the examples.
\itemize{
  \item A \dQuote{stacked} bar chart where vertical bars over length categories that each sum to 1 but are segmented by the proportion of each age in that length category is constructed with \code{type="barplot"}.  The ages will be labeled in the bar segments unless \code{showLegend=TRUE} is used.
  \item A \dQuote{stacked} are chart similar to the bar chart described above is constructed with \code{type="area"}.
  \item A plot with (different colored) lines that connect the proportions of ages within each length category is constructed with \code{type="lines"}.
  \item A plot with (different colored) lines, as estimated by loess splines, that connect the proportions of ages within each length category is constructed with \code{type="splines"}.
  \item A \dQuote{bubble} plot where circles whose size is proportional to the proportion of fish of each age in each length category is constructed with \code{type="bubble"}.  The color of the bubbles can be controlled with \code{col=} and an underlying grid for ease of seeing the age and length category for each bubble can be controlled with \code{grid=}.  Bubbles from a second age-length key can be overlaid on an already constructed bubble plot by using \code{add=TRUE} in a second call to \code{ageKeyPlot}.
}
Note that all plots are \dQuote{vertically conditional} -- i.e., each represents the proportional ages WITHIN each length category.
}
\section{fishR vignette}{
 \url{https://sites.google.com/site/fishrfiles/gnrl/AgeLengthKey.pdf}
}
\examples{
## Make an example age-length key -- same as in ageKey()
data(WR79)
WR.age <- Subset(WR79, !is.na(age))      # isolate the age sample
WR.age$LCat <- lencat(WR.age$len,w=5)    # add length categories (width=5)
raw <- xtabs(~LCat+age,data=WR.age)      # create age-length key
( WR.key <- prop.table(raw, margin=1) )

## Various visualizations of the age-length key
ageKeyPlot(WR.key,"barplot")
ageKeyPlot(WR.key,"barplot",showLegend=TRUE)
ageKeyPlot(WR.key,"area",showLegend=TRUE)
ageKeyPlot(WR.key,"splines")
ageKeyPlot(WR.key,"lines")
ageKeyPlot(WR.key,"bubble")
ageKeyPlot(WR.key,"bubble",grid=FALSE)
ageKeyPlot(WR.key,"bubble",grid="blue")
ageKeyPlot(WR.key,"bubble",grid=rgb(0,0,0,0.2),col=rgb(0,0,0,0.5))
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
See \code{\link{ageKey}} for using an age-length key to assign ages to individual fish and \code{\link{ageKeyPrep}} to \dQuote{fix} ill-formed age-length keys.
}
\keyword{plot}

