% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/agePrecision.R
\name{agePrecision}
\alias{agePrecision}
\alias{plot.agePrec}
\alias{summary.agePrec}
\title{Compute measures of precision among sets of ages.}
\usage{
agePrecision(formula, data)

\method{summary}{agePrec}(object, what = c("precision", "difference",
  "absolute difference", "detail"), percent = TRUE, digits = 4, ...)
}
\arguments{
\item{formula}{A formula of the form \code{~var1+var2+var3+...} or, alternatively, \code{var1~var2+var3+...}, where the \code{varX} generically represent the variables that contain the age assignments.  The alternative formula allows for similar code as used in \code{\link{ageBias}} and can have only one variable on the left-hand side.}

\item{data}{A data.frame that minimally contains the variables in \code{formula}.}

\item{object}{An object of class \code{agePrec}, usuall a result from \code{agePrecision}.}

\item{what}{A single string that indicates what type of summary to print.  See details.}

\item{percent}{A single logical that indicates whether the difference table (see details) should be represented as percentages (\code{TRUE}; default) or frequency (\code{FALSE}) of fish.}

\item{digits}{A single numeric that indicates the minimum number of digits to print when using \code{summary}.}

\item{\dots}{Additional arguments for methods.}
}
\value{
The main function returns a list with the following items:
\itemize{
  \item detail A data.frame with all data given in \code{data} and intermediate calculations for each fish.  See details
  \item rawdiff A frequency table of fish by differences for each pair of ages.
  \item absdiff A frequency table of fish by absolute differences for each pair of ages.
  \item APE The mean average percent error.
  \item ACV The mean coefficient of variation.
  \item n Number of fish in \code{data}.
  \item R Number of age assessments for each fish given in \code{formula}.
}

Nothing is returned by \code{summary}, but see details for what is printed.
}
\description{
Computes overall measures of precision for multiple age assignments made on the same individuals.  The age assignments may be from two or more readers of the same structure, one reader at two or more times, or two or more stuctures (e.g., scales, spines, otoliths).  Measures of precision include ACV (Average Coefficient of Variation), APE (Average Percent Error), and various percentage difference values.
}
\details{
If \code{what="precision"} in \code{summary} then a summary table that contains the following items will be printed:
\itemize{
  \item n Number of fish in \code{data}.
  \item R Number of age assessments given in \code{formula}.
  \item ACV The mean coefficient of variation.  See the fishR vignette for calculational details.
  \item APE The mean average percent error.  See the fishR vignette for calculational details.
  \item PercAgree The percentage of fish for which all age assignments perfectly agree.
}

If \code{what="difference"} is used in \code{summary} then a table that describes either the percentage (if \code{percent=TRUE}, default) or frequency of fish by the difference in paired age assignments.  This table has one row for each possible pair of age assignments.

If \code{what="absolute difference"} is used in \code{summary} then a table that describes either the percentage (if \code{percent=TRUE}, default) or frequency of fish by the absolute value of the difference in paired age assignments.  This table has one row for each possible pair of age assignments.  The \dQuote{1} column, for example, represents age assignments that disagree by one year (in either direction).

If \code{what="detail"} is used in \code{summary} then a data frame of the original \code{data} along with the intermediate caculations of the average age, standard deviation of age, APE, and ACV for each individual will be printed.  These details are generally only used to check or to understand calculations.
}
\section{Testing}{
 Tested all precision results against published results in Herbst and Marsden (2011) for the \code{\link{WhitefishLC}} data and the results for the \code{\link[FSAdata]{AlewifeLH}} data set from \pkg{FSAdata} against results from \url{http://www.nefsc.noaa.gov/fbp/age-prec/}.
}

\section{fishR vignette}{
 \url{https://sites.google.com/site/fishrfiles/gnrl/AgeComparisons.pdf}
}
\examples{
## Example with just two age assignments
data(WhitefishLC)
ap1 <- agePrecision(~otolithC+scaleC,data=WhitefishLC)
summary(ap1)
summary(ap1,what="precision")
summary(ap1,what="difference")
summary(ap1,what="difference",percent=FALSE)
summary(ap1,what="absolute",percent=FALSE)

barplot(ap1$rawdiff,ylab="Frequency",xlab="Otolith - Scale Age")
summary(ap1,what="detail")
summary

## Example with three age assignments
ap2 <- agePrecision(~otolithC+finrayC+scaleC,data=WhitefishLC)
summary(ap2)
summary(ap2,what="precision")
summary(ap2,what="difference")
summary(ap2,what="difference",percent=FALSE)
summary(ap2,what="absolute",percent=FALSE)
summary(ap2,what="detail")
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
Beamish, R.J. and D.A. Fournier.  1981.  \href{http://www.pac.dfo-mpo.gc.ca/science/people-gens/beamish/PDF_files/compareagecjfas1981.pdf}{A method for comparing the precision of a set of age determinations.}  Canadian Journal of Fisheries and Aquatic Sciences 38:982-983.

Campana, S.E.  1982.  \href{http://www.denix.osd.mil/nr/crid/Coral_Reef_Iniative_Database/References_for_Reef_Assessment_files/Campana,\%202001.pdf}{Accuracy, precision and quality control in age determination, including a review of the use and abuse of age validation methods.} Journal of Fish Biology 59:197-242.

Campana, S.E., M.C. Annand, and J.I. McMillan. 1995.  \href{http://www.bio.gc.ca/otoliths/documents/Campana\%20et\%20al\%201995\%20TAFS.pdf}{Graphical and statistical methods for determining the consistency of age determinations.} Transactions of the American Fisheries Society 124:131-138.

Chang, W.Y.B. 1982.  \href{http://www.nrcresearchpress.com/doi/abs/10.1139/f82-158}{A statistical method for evaluating the reproducibility of age determination.}  Canadian Journal of Fisheries and Aquatic Sciences 39:1208-1210.
}
\seealso{
See \code{\link{ageBias}} for computation of the full age agreement table, along with tests and plots of age bias.
}
\keyword{htest}
\keyword{manip}

