% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ALKAgeDist}
\alias{ALKAgeDist}
\title{Proportions-at-age from an Age-Length Key}
\usage{
ALKAgeDist(key, lenA.n, len.n)
}
\arguments{
\item{key}{A numeric matrix that contains the age-length key.  See details.}

\item{lenA.n}{A vector of sample sizes for each length interval in the \emph{aged sample}.}

\item{len.n}{A vector of sample sizes for each length interval in the \emph{complete sample} (i.e., all fish regardles of whether they were aged or not).}
}
\value{
A data.frame with as many rows as ages present in \code{key} and the following three variables:
\itemize{
  \item age The ages.
  \item prop The prortion of fish at each age.
  \item se The SE for the proportion of fish at each age.
 }
}
\description{
Uses the methods of Quinn and Deriso (1999) to compute the proportions-at-age (and the standard errors and coefficient of variation of those proportions) in a larger sample based on an age-length-key created from a subsample of ages through a two-stage random sampling design.
}
\details{
The age-length key sent in \code{key} must be constructed with length intervals as rows and age values as columns.  XXX
}
\section{fishR vignette}{
 none yet.
}

\section{Tests}{
 The results from this function perfectly match the results in Table 8.4 (left) of Quinn and Deriso when using \code{data(SnapperHG2)} from \pkg{FSAdata}.  The results also perfectly match the results from using \code{alkprop} in \pkg{fishmethods}.
}
\examples{
## Get data with length measurements and some assigned ages
data(WR79)

## Example -- Even breaks for length categories
WR1 <- WR79
# add length intervals (width=5)
WR1$LCat <- lencat(WR1$len,w=5)
# get number of fish in each length interval in the entire sample
len.n <- xtabs(~LCat,data=WR1)
# isolate aged sample and get number in each length interval
WR1.age <- Subset(WR1, !is.na(age))
lenA.n <- xtabs(~LCat,data=WR1.age)
# create age-length key
raw <- xtabs(~LCat+age,data=WR1.age)
( WR1.key <- prop.table(raw, margin=1) )

# use age-length key to estimate age distribution of all fish
ALKAgeDist(WR1.key,lenA.n,len.n)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
Lai, H.-N.  1987.

Lai, H.-N.  1993.

Quinn, T. J. and R. B. Deriso. 1999. Quantitative Fish Dynamics. Oxford University Press, New York, New York. 542 pages
}
\seealso{
See \code{\link{ageKey}} and related functions for a completely different methodology.  See \code{alkprop} from \pkg{fishmethods} for the exact same methodology but with different inputs.
}
\keyword{manip}

