% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ALKMeanVar}
\alias{ALKMeanVar}
\title{Mean Values-at-age from an Age-Length Key}
\usage{
ALKMeanVar(formula, data, key, lenA.n, len.n, method = c("BettoliMiranda",
  "QuinnDeriso"))
}
\arguments{
\item{formula}{A formula of the form \code{var~length+age} where \dQuote{var} generically represents the variable to be summarized (e.g., length, weight, fecundity), \dQuote{length} generically represents the variable that contains the known length measurements, and \dQuote{age} generically represents the variable that contain the assessed ages.}

\item{data}{A data.frame that minimally contains the length measurements, assessed ages, and the variable to be summarized (i.e., this should be the aged sample) as given in \code{formula}.}

\item{key}{A numeric matrix that contains the age-length key.  See details.}

\item{lenA.n}{A vector of sample sizes for each length interval in the \emph{aged sample}.}

\item{len.n}{A vector of sample sizes for each length interval in the \emph{complete sample} (i.e., all fish regardles of whether they were aged or not).}

\item{method}{A string that indicates which method of calculation should be used.  See details.}
}
\value{
A data.frame with as many rows as ages present in \code{key} and the following three variables:
\itemize{
  \item mean The mean value at each age.
  \item se The SE for the mean value at each age.
  \item cv The CV for the mean value at each age.
 }
}
\description{
Uses the methods of Quinn and Deriso (1999) to compute the mean value-at-age (and the standard errors and coefficient of variation of those means) in a larger sample based on an age-length-key created from a subsample of ages through a two-stage random sampling design.  The mean values could be mean length-, weight-, or fecundity-at-age, for example.
}
\details{
The age-length key sent in \code{key} must be constructed with length intervals as rows and age values as columns.  XXX
}
\section{fishR vignette}{
 none yet.
}
\examples{
## Get data with length measurements and some assigned ages
data(WR79)

## Example -- Even breaks for length categories
WR1 <- WR79
# add length intervals (width=5)
WR1$LCat <- lencat(WR1$len,w=5)
# get number of fish in each length interval in the entire sample
len.n <- xtabs(~LCat,data=WR1)
# isolate aged sample and get number in each length interval
WR1.age <- Subset(WR1, !is.na(age))
lenA.n <- xtabs(~LCat,data=WR1.age)
# create age-length key
raw <- xtabs(~LCat+age,data=WR1.age)
( WR1.key <- prop.table(raw, margin=1) )

## use age-length key to estimate mean length-at-age of all fish
# Bettoli-Miranda method
ALKMeanVar(len~LCat+age,WR1.age,WR1.key,lenA.n,len.n)

# Quinn-Deriso method
ALKMeanVar(len~LCat+age,WR1.age,WR1.key,lenA.n,len.n,method="QuinnDeriso")
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
Bettoli,

Quinn, T. J. and R. B. Deriso. 1999. Quantitative Fish Dynamics. Oxford University Press, New York, New York. 542 pages
}
\seealso{
See \code{\link{ageKey}} and related functions for a completely different methodology.  See \code{\link{ALKAgeDist}} for a related method of determining the proportion of fish at each age.
}
\keyword{manip}

