% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BluegillJL.R
\docType{data}
\name{BluegillJL}
\alias{BluegillJL}
\title{Capture histories (2 samples) of Bluegill from Jewett Lake, MI.}
\format{A data frame with 277 observations on the following 2 variables.
 \describe{
   \item{first}{a numeric vector of indicator variables for the first sample (1=captured)}
   \item{second}{a numeric vector of indicator variables for the second sample (1=captured)}
 }}
\source{
From example 8.1 in Schneider, J.C. 1998. Lake fish population estimates by mark-and-recapture methods. Chapter 8 in Schneider, J.C. (ed.) 2000. Manual of fisheries survey methods II: with periodic updates. Michigan Department of Natural Resources, Fisheries Special Report 25, Ann Arbor. [Was (is?) from http://www.michigandnr.com/publications/pdfs/IFR/manual/SMII\%20Chapter08.pdf.]
}
\description{
Each line consists of the capture history over two samples of Bluegill (\emph{Lepomis macrochirus}) in Jewett Lake (MI). This file contains the capture histories for only Bluegill larger than 6-in.
}
\section{Topic(s)}{

 \itemize{
   \item Population Size
   \item Abundance
   \item Mark-Recapture
   \item Capture-Recapture
   \item Petersen
   \item Capture History
 }
}

\examples{
str(BluegillJL)
head(BluegillJL)

}
\seealso{
Used in \code{\link{mrClosed}} examples.
}
\concept{Abundance}
\concept{Capture History}
\concept{Capture-Recapture}
\concept{Mark-Recapture}
\concept{Petersen}
\concept{Population Size}
\keyword{datasets}
