% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{predict.boot}
\alias{predict.boot}
\alias{htest.boot}
\alias{plot.boot}
\alias{htest.boot}
\alias{plot.boot}
\title{Associated S3 methods for Boot from car.}
\usage{
\method{predict}{boot}(object, FUN, conf.level = 0.95, level = conf.level,
  digits = NULL, ...)

\method{htest}{boot}(object, parm = NULL, bo = 0, alt = c("two.sided",
  "less", "greater"), plot = FALSE, ...)

\method{plot}{boot}(x, ...)
}
\arguments{
\item{object, x}{A \code{boot} object returned from \code{\link[car]{Boot}} from the \pkg{car} package.}

\item{FUN}{The function to be applied for the prediction. See the examples.}

\item{conf.level, level}{A level of confidence as a proportion.}

\item{digits}{A single numeric that indicates the number of digits for the result.}

\item{\dots}{Additional items to send to functions. See details.}

\item{parm}{A number or string that indicates which column of \code{object} contains the parameter estimates to use for hypothesis test.}

\item{bo}{The null hypothesized parameter value.}

\item{alt}{A string that indicates the \dQuote{direction} of the alternative hypothesis. See details.}

\item{plot}{A logical that indicates whether a histogram of the \code{parm} parameters with a vertical line illustrating the \code{bo} value will be constructed with \code{htest}.}
}
\value{
\code{htest} returns a two-column matrix with the first column containing the hypothesized value sent to this function and the second column containing the corresponding p-value.

\code{plot} constructs scatterplots of all pairs of bootstrapped parameter estimates.

\code{predict} returns a matrix with one row and three columns, with the first column holding the predicted value (i.e., the median prediction) and the last two columns holding the approximate confidence interval.
}
\description{
Provides S3 methods to construct non-parametric bootstrap, hypothesis tests, predictions, and plots of the parameter estimates for \code{boot} objects returned from \code{\link[car]{Boot}} from the \pkg{car} package.
}
\details{
See \code{\link[car]{confint.boot}} in \pkg{car} for methods to construct confidence intervals from the bootstrapped results.

\code{predict} applies a user-supplied function to each row of \code{object} and then finds the median and the two quantiles that have the proportion (1-\code{conf.level})/2 of the bootstrapped predictions below and above. The median is returned as the predicted value and the quantiles are returned as an approximate 100\code{conf.level}\% confidence interval for that prediction. Values for the independent variable in \code{FUN} must be a named argument sent in the \dots argument (see examples). Note that if other arguments are needed in \code{FUN} besides values for the independent variable, then these are included in the \dots argument AFTER the values for the independent variable.

In \code{htest} the \dQuote{direction} of the alternative hypothesis is identified by a string in the \code{alt=} argument. The strings may be \code{"less"} for a \dQuote{less than} alternative, \code{"greater"} for a \dQuote{greater than} alternative, or \code{"two.sided"} for a \dQuote{not equals} alternative (the DEFAULT). In the one-tailed alternatives the p-value is the proportion of bootstrapped parameter estimates in \code{object$coefboot} that are extreme of the null hypothesized parameter value in \code{bo}. In the two-tailed alternative the p-value is twice the smallest of the proportion of bootstrapped parameter estimates above or below the null hypothesized parameter value in \code{bo}.
}
\examples{
data(Ecoli)
fnx <- function(days,B1,B2,B3) {
  if (length(B1) > 1) {
    B2 <- B1[2]
    B3 <- B1[3]
    B1 <- B1[1]
  }
  B1/(1+exp(B2+B3*days))
}
nl1 <- nls(cells~fnx(days,B1,B2,B3),data=Ecoli,
           start=list(B1=6,B2=7.2,B3=-1.45))
if (require(car)) {    # for Boot()
  nl1.boot <- car::Boot(nl1,R=99)  # B=99 too small to be useful
  confint(nl1.boot,"B1")
  confint(nl1.boot,c(2,3))
  confint(nl1.boot,level=0.90)
  predict(nl1.boot,fnx,days=1:3)
  predict(nl1.boot,fnx,days=3)
  htest(nl1.boot,1,bo=6,alt="less")
  hist(nl1.boot)
  plot(nl1.boot)
  cor(nl1.boot$t)
}

}
\references{
S. Weisberg (2005). \emph{Applied Linear Regression}, third edition. New York: Wiley, Chapters 4 and 11.
}
\seealso{
\code{\link[car]{Boot}}, \code{confint.boot}, and \code{hist.boot} in \pkg{car}.
}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
\keyword{htest}
