% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bootCase.R
\name{confint.bootCase}
\alias{confint.bootCase}
\alias{hist.bootCase}
\alias{htest.bootCase}
\alias{plot.bootCase}
\alias{predict.bootCase}
\title{Associated S3 methods for bootCase from car.}
\usage{
\method{confint}{bootCase}(object, parm = NULL, level = conf.level,
  conf.level = 0.95, ...)

\method{htest}{bootCase}(object, parm = NULL, bo = 0, alt = c("two.sided",
  "less", "greater"), ...)

\method{hist}{bootCase}(x, same.ylim = TRUE, ymax = NULL, col = "gray90",
  nr = round(sqrt(ncol(x))), nc = ceiling(sqrt(ncol(x))), right = FALSE,
  ...)

\method{plot}{bootCase}(x, nr = round(sqrt(np)), nc = ceiling(sqrt(np)),
  ...)

\method{predict}{bootCase}(object, FUN, MARGIN = 1, conf.level = 0.95,
  digits = NULL, ...)
}
\arguments{
\item{object}{A \code{bootCase} matrix object for \code{confint} or \code{htest}.}

\item{parm}{A number that indicates which column of \code{object} contains the parameter estimates to use in the hypothesis test.}

\item{level}{Same as \code{conf.level}.}

\item{conf.level}{A level of confidence as a proportion.}

\item{bo}{The null hypothesized parameter value.}

\item{alt}{A string that indicates the \dQuote{direction} of the alternative hypothesis.  See details.}

\item{x}{A \code{bootCase} object for \code{hist}, \code{plot} or \code{boxplot}.}

\item{same.ylim}{A logicial that indicates whether the same limits for the y-axis should be used on each histogram.  Defaults to \code{TRUE}.  Ignored if \code{ylmts} is non-null.}

\item{ymax}{A single value that sets the maximum y-axis limit for each histogram or a vector of length equal to the number of groups that sets the maximum y-axis limit for each histogram separately.}

\item{col}{A named color for the histogram bars.}

\item{nr}{A numeric that contains the number of rows to use on the graphic.}

\item{nc}{A numeric that contains the number of columns to use on the graphic.}

\item{right}{A logical that indicates if the histogram bins are right-closed (left open) intervals (\code{=TRUE}) or not (\code{=FALSE}; default).}

\item{FUN}{The function to be applied.}

\item{MARGIN}{A single numeric that indicates the margin over which to apply the function.  \code{MARGIN=1} will apply to each row and is the default.}

\item{digits}{A single numeric that indicates the number of digits for the result.}

\item{\dots}{Additional items to send to functions.}
}
\value{
\code{confint} will return a matrix with as many rows as columns (i.e., parameter estimates) in \code{object} and two columns of the quantiles that correspond to the approximate confidence interval if \code{object} is a matrix.  If \code{object} is a vector then \code{confint} will return a vector with the two quantiles that correspond to the approximate confidence interval.

\code{htest} returns a two-column matrix with the first column containing the hypothesized value sent to this function and the second column containing the corresponding p-value.

\code{hist} constructs histograms of the bootstrapped parameter estimates.

\code{plot} constructs scatterplots of all pairs of bootstrapped parameter estimates.

\code{predict} applies a user-supplied function to each row of \code{object} and then finds the median and the two quantiles that have the proportion (1-\code{conf.level})/2 of the bootstrapped predictions below and above.  The median is returned as the predicted value and the quantiles are returned as an approximate 100\code{conf.level}\% confidence interval for that prediction.
}
\description{
Provides S3 methods to construct non-parametric bootstrap confidence intervals, hypothesis tests, and plots of the parameter estimates for a \code{\link[car]{bootCase}} object (from the \pkg{car} package).
}
\details{
\code{confint} finds the two quantiles that have the (1-\code{conf.level})/2 proportion of bootstrapped parameter estimates below and above.  This is an approximate 100\code{conf.level}\% confidence interval.

In the one-tailed alternatives, the p-value is the proportion of bootstrapped parameter estimates in that are extreme of the null hypothesized parameter value (\code{bo}).  In the two-tailed alternative, the p-value is twice the smallest of the proportion of bootstrapped parameter \code{bo}.
}
\examples{
require(car)    # for bootCase()
data(Ecoli)
fnx <- function(days,B1,B2,B3) {
  if (length(B1) > 1) {
    B2 <- B1[2]
    B3 <- B1[3]
    B1 <- B1[1]
  }
  B1/(1+exp(B2+B3*days))
}
nl1 <- nls(cells~fnx(days,B1,B2,B3),data=Ecoli,start=list(B1=6,B2=7.2,B3=-1.45))
nl1.boot <- bootCase(nl1,B=99)  # B=99 too small to be useful
confint(nl1.boot,"B1")
confint(nl1.boot,c(2,3))
confint(nl1.boot,conf.level=0.90)
htest(nl1.boot,1,bo=6,alt="less")
hist(nl1.boot)
plot(nl1.boot)
cor(nl1.boot)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
S. Weisberg (2005). \emph{Applied Linear Regression}, third edition.  New York: Wiley, Chapters 4 and 11.
}
\seealso{
\code{\link[car]{bootCase}} in \pkg{car}.
}
\keyword{htest}

