% Generated by roxygen2 (4.0.1): do not edit by hand
\name{capHistConvert}
\alias{capHistConvert}
\title{Convert between capture history data.frame formats.}
\usage{
capHistConvert(df, cols2ignore = NULL, in.type = c("frequency", "event",
  "individual", "MARK", "marked", "RMark"), out.type = c("individual",
  "event", "frequency", "MARK", "marked", "RMark"), id = NULL,
  event.ord = NULL, freq = NULL, var.lbls = NULL,
  var.lbls.pre = "event", include.id = ifelse(is.null(id), FALSE, TRUE))
}
\arguments{
\item{df}{A data.frame that contains the capture histories and, perhaps, a unique fish identifier or frequency variable.  See details.}

\item{cols2ignore}{A string or numeric vector that indicates columns in \code{df} to ignore.  Typical columns to ignore are those that are not either in \code{id=} or \code{freq=} or part of the capture history data).}

\item{in.type}{A single string that indicates the type of capture history format to convert \bold{FROM}.}

\item{out.type}{A single string that indicates the type of capture history format to convert \bold{TO}.}

\item{id}{A string or numeric that indicates the column in \code{df} that contains the unique identifier for an individual fish.  This argument is only used if \code{in.type="event"}, \code{in.type="individual"}, or, possibly, \code{in.type="RMark"}.}

\item{freq}{A string or numeric that indicates the column in \code{df} that contains the frequency of individual fish corresponding to a capture history.  This argument is only used if \code{in.type="MARK"}, \code{in.type="frequency"}, or, possibly, \code{in.type="RMark"}.}

\item{event.ord}{A string that contains a vector of ordered levels to be used when \code{in.type="event"}.  The default is to order alphabetically which may not be desirable if the events are labelled as \sQuote{first}, \sQuote{second}, \sQuote{third}, and \sQuote{fourth}.  In this case, use \code{event.ord=c("first","second","third","fourth")}.}

\item{var.lbls}{A string vector of labels for the columns that contain the returned individual or frequency capture histories.  If \code{var.lbls=NULL} or the length is different then the number of events then default labels using \code{var.lbls.pre} will be used.  This argument is only used if \code{out.type="frequency"} or \code{out.type="individual"}.}

\item{var.lbls.pre}{A single string used as a prefix for the labels of the columns that contain the returned individual or frequency capture histories.  This prefix will be appended with a number corresponding to the sample event.  This argument is only used if \code{out.type="frequency"} or \code{out.type="individual"} and will be ignored if a proper vector is given in \code{var.lbls}.}

\item{include.id}{A logical that indicates whether a unique fish identifier variable/column should be included in the output data.frame.  This argument is only used if \code{out.type="individual"} or \code{out.type="RMark"}.}
}
\value{
A data frame of the proper type given in \code{out.type} is returned.  See details.
}
\description{
Use to convert between simple versions of several capture history data.frame formats -- \dQuote{individual}, \dQuote{frequency}, \dQuote{event}, \dQuote{MARK}, and \dQuote{RMark}.  The primary use is to convert to the \dQuote{individual} format for use in \code{\link{capHistSum}}.
}
\details{
\code{\link{capHistSum}} requires capture histories to be recorded in the \dQuote{individual} format.  In this format, the data frame contains (at least) as many columns as sample events and as many rows as individually tagged fish.  Optionally, the data.frame may also contain a column with unique fish identifiers (e.g., tag numbers).  Each cell in the capture history portion of the data.frame contains a \sQuote{0} if the fish of that row was NOT seen in the event of that column and a \sQuote{1} if the fish of that row WAS seen in the event of that column.  For example, suppose that five fish were marked on four sampling events; fish \sQuote{17} was captured on the first two events; fish \sQuote{18} was captured on the first and third events; fish \sQuote{19} was captured on only the third event; fish \sQuote{20} was captured on only the fourth event; and fish \sQuote{21} was captured on the first and second events.  The \dQuote{individual} capture history date.frame for these data looks like:

\tabular{ccccc}{
fish \tab event1 \tab event2 \tab event3 \tab event4 \cr
17 \tab 1 \tab 1 \tab 0 \tab 0 \cr
18 \tab 1 \tab 0 \tab 1 \tab 0 \cr
19 \tab 0 \tab 0 \tab 1 \tab 0 \cr
20 \tab 0 \tab 0 \tab 0 \tab 1 \cr
21 \tab 1 \tab 1 \tab 0 \tab 0 \cr
}

The \dQuote{frequency} format data.frame (this format is used in \pkg{rcapture}) has unique capture histories in separate columns, as in the \dQuote{individual} format, but also includes a column with the frequency of individuals that had the capture history of that row.  It will not contain a fish identifier variable.  The same data from above looks like:

\tabular{ccccc}{
event1 \tab event2 \tab event3 \tab event4 \tab freq \cr
1 \tab 1 \tab 0 \tab 0 \tab 2 \cr
1 \tab 0 \tab 1 \tab 0 \tab 1 \cr
0 \tab 0 \tab 1 \tab 0 \tab 1 \cr
0 \tab 0 \tab 0 \tab 1 \tab 1 \cr
}

The \dQuote{event} format data.framed has a column with the unique fish identifier and a column with the event in which the fish of that row was observed.  The same data from above looks like:

\tabular{cc}{
fish \tab event \cr
17 \tab 1 \cr
18 \tab 1 \cr
21 \tab 1 \cr
17 \tab 2 \cr
21 \tab 2 \cr
18 \tab 3 \cr
19 \tab 3 \cr
20 \tab 4 \cr
}

MARK is the \dQuote{gold-standard} software for analyzing complex capture history information.  In the \dQuote{MARK} format the 0s and 1s of the capture histories are combined together as a string without any spaces and ended with a semicolon.  Thus, the \dQuote{MARK} format has the capture history strings in one column with an additional column that contains the frequency of individuals that exhibited the capture history of that row.  The same data from above looks like:

\tabular{cc}{
ch \tab freq \cr
0001; \tab 1 \cr
0010; \tab 1 \cr
1010; \tab 1 \cr
1100; \tab 2 \cr
}

\pkg{RMark} and \pkg{marked} are R packages used to replace or interact with MARK.  The \dQuote{RMark} or \dQuote{marked} format requires the capture histories as one string (must be a character string and called \sQuote{ch}), as in the \dQuote{MARK} format, but without the semicolon.  The data.frame may be augmented with an indentifier for individual fish OR with a frequency variable.  If augmented with a unique fish identification variable then the same data from above looks like:

\tabular{cc}{
fish \tab ch \cr
17 \tab 1100 \cr
18 \tab 1010 \cr
19 \tab 0010 \cr
20 \tab 0001 \cr
21 \tab 1100 \cr
}

However, if augmented with a frequency variable then the same data from above looks like:

\tabular{cc}{
ch \tab freq \cr
0001 \tab 1 \cr
0010 \tab 1 \cr
1010 \tab 1 \cr
1100 \tab 2 \cr
}

Each of the formats can be used to convert from (i.e., in \code{in.type=}) or to convert to (i.e., in \code{out.type=}) with the exception that only the individual fish identifier version can be converted to when \code{out.type="RMark"}.
}
\note{
The formats as used here are simple in the sense that one is only allowed to have the individual fish identifier or the frequency variable in addition to the capture history information.  More complex analyses may use a number of covariates.  For these more complex analyses, one should work directly with the \pkg{rcapture}, \pkg{RMark}, or \pkg{marked} packages.

This function also assumes that all unmarked captured fish are marked and returned to the population (i.e., no losses at the time of marking are allowed).
}
\section{fishR vignette}{
 \url{https://sites.google.com/site/fishrfiles/gnrl/MRClosed.pdf}, \url{https://sites.google.com/site/fishrfiles/gnrl/MROpen.pdf}
}
\examples{
## A small example of 'event' format
( ex1 <- data.frame(fish=c(17,18,21,17,21,18,19,20),yr=c(1987,1987,1987,1988,1988,1989,1989,1990)) )
# convert to 'individual' format
( ex1.E2I <- capHistConvert(ex1,id="fish",in.type="event") )
# convert to 'frequency' format
( ex1.E2F <- capHistConvert(ex1,id="fish",in.type="event",out.type="frequency") )
# convert to 'MARK' format
( ex1.E2M <- capHistConvert(ex1,id="fish",in.type="event",out.type="MARK") )
# convert to 'RMark' format
( ex1.E2R <- capHistConvert(ex1,id="fish",in.type="event",out.type="RMark") )

## convert converted 'individual' format ...
# to 'frequency' format (must ignore "id")
( ex1.I2F <- capHistConvert(ex1.E2I,id="fish",in.type="individual",out.type="frequency") )
# to 'MARK' format
( ex1.I2M <- capHistConvert(ex1.E2I,id="fish",in.type="individual",out.type="MARK") )
# to 'RMark' format
( ex1.I2R <- capHistConvert(ex1.E2I,id="fish",in.type="individual",out.type="RMark") )
# to 'event' format
( ex1.I2E <- capHistConvert(ex1.E2I,id="fish",in.type="individual",out.type="event") )

#' ## convert converted 'frequency' format ...
# to 'individual' format
( ex1.F2I <- capHistConvert(ex1.E2F,freq="freq",in.type="frequency") )
# to 'Mark' format
( ex1.F2M <- capHistConvert(ex1.E2F,freq="freq",in.type="frequency",out.type="MARK") )
# to 'RMark' format
( ex1.F2R <- capHistConvert(ex1.E2F,freq="freq",in.type="frequency",out.type="RMark") )
# to 'event' format
( ex1.F2E <- capHistConvert(ex1.E2F,freq="freq",in.type="frequency",out.type="event") )

## convert converted 'MARK' format ...
# to 'individual' format
( ex1.M2I <- capHistConvert(ex1.E2M,freq="freq",in.type="MARK") )
# to 'frequency' format
( ex1.M2F <- capHistConvert(ex1.E2M,freq="freq",in.type="MARK",out.type="frequency") )
# to 'RMark' format
( ex1.M2R <- capHistConvert(ex1.E2M,freq="freq",in.type="MARK",out.type="RMark") )
# to 'event' format
( ex1.M2E <- capHistConvert(ex1.E2M,freq="freq",in.type="MARK",out.type="event") )

## convert converted 'RMark' format ...
# to 'individual' format
( ex1.R2I <- capHistConvert(ex1.E2R,id="fish",in.type="RMark") )
# to 'frequency' format
( ex1.R2F <- capHistConvert(ex1.E2R,id="fish",in.type="RMark",out.type="frequency") )
# to 'MARK' format
( ex1.R2M <- capHistConvert(ex1.E2R,id="fish",in.type="RMark",out.type="MARK") )
# to 'event' format
( ex1.R2E <- capHistConvert(ex1.E2R,id="fish",in.type="RMark",out.type="event") )

## Remove semi-colon from MARK format to make a RMark 'frequency' format
ex1.E2R1 <- ex1.E2M
ex1.E2R1$ch <- sub(";","",ex1.E2R1$ch)
ex1.E2R1
# convert this to 'individual' format
( ex1.R2I1 <- capHistConvert(ex1.E2R1,freq="freq",in.type="RMark") )
# convert this to 'frequency' format
( ex1.R2F1 <- capHistConvert(ex1.E2R1,freq="freq",in.type="RMark",out.type="frequency") )
# convert this to 'MARK' format
( ex1.R2M1 <- capHistConvert(ex1.E2R1,freq="freq",in.type="RMark",out.type="MARK") )
# convert this to 'event' format
( ex1.R2E1 <- capHistConvert(ex1.E2R1,freq="freq",in.type="RMark",out.type="event") )


#' ## A small example using character ids
( ex2 <- data.frame(fish=c("id17","id18","id21","id17","id21","id18","id19","id20"),
                    yr=c(1987,1987,1987,1988,1988,1989,1989,1990)) )
# convert to 'individual' format
( ex2.E2I <- capHistConvert(ex2,id="fish",in.type="event") )
# convert to 'frequency' format
( ex2.E2F <- capHistConvert(ex2,id="fish",in.type="event",out.type="frequency") )
# convert to 'MARK' format
( ex2.E2M <- capHistConvert(ex2,id="fish",in.type="event",out.type="MARK") )
# convert to 'RMark' format
( ex2.E2R <- capHistConvert(ex2,id="fish",in.type="event",out.type="RMark") )

## convert converted 'individual' format ...
# to 'frequency' format (must ignore "id")
( ex2.I2F <- capHistConvert(ex2.E2I,id="fish",in.type="individual",out.type="frequency") )
# to 'MARK' format
( ex2.I2M <- capHistConvert(ex2.E2I,id="fish",in.type="individual",out.type="MARK") )
# to 'RMark' format
( ex2.I2R <- capHistConvert(ex2.E2I,id="fish",in.type="individual",out.type="RMark") )
# to 'event' format
( ex2.I2E <- capHistConvert(ex2.E2I,id="fish",in.type="individual",out.type="event") )


## ONLY RUN IN INTERACTIVE MODE
if (interactive()) {

## A larger example of 'frequency' format
require(Rcapture)
data(bunting)
head(bunting)
# convert to 'individual' format
ex2a <- capHistConvert(bunting,in.type="frequency",freq="freq")
head(ex2a)
# convert to 'MARK' format
ex2b <- capHistConvert(bunting,id="id",in.type="frequency",freq="freq",out.type="MARK")
head(ex2b)
# convert converted 'individual' back to 'MARK' format
ex2c <- capHistConvert(ex2a,id="id",in.type="individual",out.type="MARK")
head(ex2c)
# convert converted 'individual' back to 'frequency' format
ex2d <- capHistConvert(ex2a,id="id",in.type="individual",out.type="frequency",var.lbls.pre="Sample")
head(ex2d)

}


## A small example of 'MARK' format -- capture history followed by frequency
( ex3 <- data.frame(ch=c("10101;","10001;","01010;"),freq=c(3,1,2)) )
# convert to 'individual' format
( ex3a <- capHistConvert(ex3,in.type="MARK",freq="freq") )
# convert to 'frequency' format
( ex3b <- capHistConvert(ex3,in.type="MARK",freq="freq",out.type="frequency") )


## A small example of 'MARK' format with two groups -- males and females
( ex4 <- data.frame(ch=c("100101;","010001;"),male=c(3,1),female=c(2,2)) )
# convert to 'individual' format
( ex4m <- capHistConvert(ex4[,c("ch","male")],in.type="MARK",freq="male",include.id=FALSE) )
( ex4f <- capHistConvert(ex4[,c("ch","female")],in.type="MARK",freq="female",include.id=FALSE) )
require(gdata)   # for combine()
( ex4a <- combine(ex4m,ex4f,names=c("male","female")) )
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
\code{\link{capHistSum}} to summarize \dQuote{individual} capture histories into a format useable in \code{\link{mrClosed}} and \code{\link{mrOpen}}.  Also see \pkg{rcapture}, \pkg{RMark}, or \pkg{marked} packages.
}
\keyword{manip}

