% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/confint.nlsBoot.R
\name{confint.nlsBoot}
\alias{confint.nlsBoot}
\alias{confint.nlsboot}
\title{Construct a confidence interval from nlsBoot results.}
\usage{
\method{confint}{nlsBoot}(object, parm = NULL, level = 0.95,
  conf.level = level, plot = FALSE, err.col = "red", err.lwd = 2,
  rows = round(sqrt(ncol(object$coefboot))),
  cols = ceiling(sqrt(ncol(object$coefboot))), ...)
}
\arguments{
\item{object}{An object saved from \code{nlsBoot()}.}

\item{parm}{An integer that indicates which parameter to compute the confidence interval for.  Will compute for all parameters if \code{NULL}.}

\item{level}{Same as \code{conf.level}.  Used for compatability with the main \code{confint}.}

\item{conf.level}{A level of confidence as a proportion.}

\item{plot}{A logical that indicates whether a histogram of the \code{parm} parameters from the bootstrap samples with error bars that illustrate the bootstrapped confidence intervals should be constructed.}

\item{err.col}{A single numeric or character that identifies the color for the error bar on the plot.}

\item{err.lwd}{A single numeric that identifies the line width for the error bar on the plot.}

\item{rows}{A numeric that contains the number of rows to use on the graphic.}

\item{cols}{A numeric that contains the number of columns to use on the graphic.}

\item{\dots}{Additional arguments for \code{\link{hist.formula}} if \code{plot=TRUE}.}
}
\value{
A matrix with as many rows as columns (i.e., parameter estimates) in the \code{object$coefboot} data frame and two columns of the quantiles that correspond to the approximate confidence interval.
}
\description{
Constructs a non-parametric bootstrap confidence interval from an object of \code{nlsBoot()} (in the \pkg{nlstools} package) results.
}
\details{
This function finds the two quantiles that have the proportion (1-\code{conf.level})/2 of the bootstrapped parameter estimates below and above.  This is an approximate 100\code{conf.level}\% confidence interval.
}
\examples{
## ONLY RUN IN INTERACTIVE MODE
if (interactive()) {

require(nlstools)
example(nlsBoot)
confint(boo,conf.level=0.90)
confint(boo,conf.level=0.90,plot=TRUE)

} ## END IF INTERACTIVE MODE
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
\code{\link{predict.nlsBoot}}, \code{summary.nlsBoot} in \pkg{nlstools}
}
\keyword{htest}

