% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dunnTest.R
\name{dunnTest}
\alias{dunnTest}
\alias{dunnTest.default}
\alias{dunnTest.formula}
\alias{print.dunnTest}
\title{Dunn's Kruskal-Wallis Multiple Comparisons.}
\usage{
dunnTest(x, ...)

\method{dunnTest}{default}(x, g, method = stats::p.adjust.methods[c(4, 1:3,
  5:8)], ...)

\method{dunnTest}{formula}(x, data = NULL,
  method = stats::p.adjust.methods[c(4, 1:3, 5:8)], ...)

\method{print}{dunnTest}(x, ...)
}
\arguments{
\item{x}{A numeric vector of data values or a formula of the form x~g.}

\item{g}{A factor vector or a (non-numeric) vector that can be coerced to a factor vector.}

\item{method}{A single string that identifies the method to use to control the experimentwise error rate.  See details in \code{\link[stats]{p.adjust}}.}

\item{data}{A data.frame that minimally contains \code{x} and \code{g}.}

\item{\dots}{Not yet used.}
}
\value{
A list with two items -- \code{method} is the long name of the method used to control the experimentwise error rate and a data.frame in \code{res} with the following variables:
\itemize{
  \item Comparison Labels for each pairwise comparison.
  \item Z Values for the Z test statistic for each comparison.
  \item P.unadj Unadjusted p-values for each comparison.
  \item P.adj Adjusted p-values for each comparison.
}
}
\description{
Performs Dunn's (1964) test of multiple comparisons following a significant Kruskal-Wallis test, possibly with a correction to control the experimentwise error rate.
}
\details{
This function performs \dQuote{Dunn's} test of multiple comparisons following a Kruskal-Wallis test.  Unadjusted two-sided p-values for each pairwise comparison among groups are computed following Dunn's description.  These p-values may be adjusted with the methods of \code{\link[stats]{p.adjust}}.

Dunn's method can be described in the following way.  To compare groups i and j, the absolute value of the difference between the mean rank of group i and the mean rank of group j is found.  If there are no ties, this difference in mean ranks is divided by the square root of [(N*(N+1)/12)*((1/Ni)+(1/Nj))], where N is the total number of individuals in all groups, and ni and nj are the number of individuals in the groups i and j, respectively. If there are ties, the difference in mean ranks is divided by the square root of [(N*(N+1)-Sum((ti^3)-ti)/(N-1))/12*((1/ni)+(1/nj))], where ti is the number of ties in the ith group of ties.  These ratios are a Z test statistic and the two-sided p-value is computed as 2*Pr(Z>|z|).
}
\note{
The data.frame will be reduced to only those rows that are complete cases for \code{x} and \code{g}.  In other words, rows with missing data for either \code{x} or \code{g} are removed from the analysis.

There are a number of functions in other packages that purport to do similar analyses.

The \code{\link[dunn.test]{dunn.test}} function in \pkg{dunn.test} performs the same calculations but does not use formula notation, does not produce two-sided p-values, and prints results automatically and in a format that I consider cumbersome.  It does, however, provide a few more methods for controlling the experimentwise error rate.  The results from \code{dunnTest} (here) match those from \code{\link[dunn.test]{dunn.test}} (if the p-values are adjusted to be two-sided) for the Bonferroni method and for some but not all p-values for the Holm, Hochberg, Benjamini-Hochberg, and Benjamini-Yekuteili methods.  Most differences are very slight.  The author of \code{\link[dunn.test]{dunn.test}} says that the difference is due to \code{\link[dunn.test]{dunn.test}} utilizing the order of the p-values where \code{\link[stats]{p.adjust}} (used in \code{dunnTest} (here)) does not.

The \code{\link[asbio]{pairw.kw}} function from the \pkg{asbio} package performs the Dunn test with the Bonferroni correction.  The p-value results from \code{DunnTest} match the results from \code{\link[asbio]{pairw.kw}}.  The \code{\link[asbio]{pairw.kw}} also provides a confidence interval for the difference in mean ranks between pairs of groups.

The \code{\link[PMCMR]{posthoc.kruskal.nemenyi.test}} function from the \pkg{PMCMR} package uses the \dQuote{Nemenyi} (1963) method of multiple comparisons.

The \code{\link[pgirmess]{kruskalmc}} function from the \pkg{pgirmess} package uses the method described by Siegel and Castellan (1988).

It is not clear which method the \code{\link[agricolae]{kruskal}} function from the \pkg{agricolae} package uses.  It does not seem to output p-values but it does allow for a wide variety of methods to control the experimentise error rate.
}
\examples{
## pH in four ponds data from Zar (2010)
ponds <- data.frame(pond=as.factor(rep(1:4,each=8)),
                    pH=c(7.68,7.69,7.70,7.70,7.72,7.73,7.73,7.76,
                         7.71,7.73,7.74,7.74,7.78,7.78,7.80,7.81,
                         7.74,7.75,7.77,7.78,7.80,7.81,7.84,NA,
                         7.71,7.71,7.74,7.79,7.81,7.85,7.87,7.91))
ponds <- ponds[complete.cases(ponds),]

# non-formula usage (default "bonferroni" method)
dunnTest(ponds$pH,ponds$pond)

# formula usage (default "Bonferroni" method)
dunnTest(pH~pond,data=ponds)

# other methods
dunnTest(pH~pond,data=ponds,method="holm")
dunnTest(pH~pond,data=ponds,method="BH")
dunnTest(pH~pond,data=ponds,method="none")
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}, but this is largely simple modifications (return result, two-sided p-values, and use of \code{\link[stats]{p.adjust}}) of \code{\link[dunn.test]{dunn.test}} in \pkg{dunn.test}.
}
\references{
Dunn, O.J. 1964. Multiple comparisons using rank sums. Technometrics 6:241-252.
}
\seealso{
See \code{\link{kruskal.test}} and \code{\link[dunn.test]{dunn.test}} in \pkg{dunn.test}, \code{\link[PMCMR]{posthoc.kruskal.nemenyi.test}} in \pkg{PMCMR}, \code{\link[pgirmess]{kruskalmc}} in \pkg{pgirmess}, and \code{\link[agricolae]{kruskal}} in \pkg{agricolae}.
}

