% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSAUtils.R
\name{fact2num}
\alias{fact2num}
\title{Converts "numeric" factor levels to numeric values.}
\usage{
fact2num(object)
}
\arguments{
\item{object}{A vector with \dQuote{numeric} factor levels to be converted to numeric values.}
}
\value{
A numeric vector.
}
\description{
Converts \dQuote{numeric} factor levels to numeric values.
}
\examples{
junk <- factor(c(1,7,2,4,3,10))
str(junk)
junk2 <- fact2num(junk)
str(junk2)

## ONLY RUN IN INTERACTIVE MODE
if (interactive()) {

bad <- factor(c("A","B","C"))
# This will result in an error -- levels are not 'numeric'
bad2 <- fact2num(bad)

}  ## END IF INTERACTIVE MODE

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
\keyword{manip}
