% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterd.R
\name{filterD}
\alias{filterD}
\title{Filters a data frame and drops the unused levels.}
\usage{
filterD(.data, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{\dots}{further arguments to be passed to \code{\link[dplyr]{filter}} from \pkg{dplyr}.}
}
\value{
A data frame with same variables as \code{.data}.
}
\description{
Filters a data frame and drops the unused levels.
}
\details{
This function is equivalent to the combined usage of \code{\link[dplyr]{filter}} from \pkg{dplyr} and \code{\link{droplevels}}.
}
\note{
Newbie students using R expect that when a data.frame is filtered via a factor variable that any original levels that are no longer used after the filterting will be ignored.  This, however, is not the case and often results in tables with empty cells and figures with empty bars.  One remedy is to use \code{\link{droplevels}} immediately following \code{\link[dplyr]{filter}}.  This generally becomes a repetitive sequence; thus, \code{filterD} incorporates these two functions into one function.
}
\section{IFAR Chapter}{
 \href{https://fishr.wordpress.com/books/ifar/}{2-Basic Data Manipulations}.
}
\examples{
## The problem -- note use of unused level in the final table.
require(dplyr)
levels(iris$Species)
tmp1 <- dplyr::filter(iris,Species=="setosa" | Species=="versicolor")
levels(tmp1$Species)
xtabs(~Species,data=tmp1)

## A simpler fix using filterD
tmp2 <- filterD(iris,Species=="setosa" | Species=="versicolor")
levels(tmp2$Species)
xtabs(~Species,data=tmp2)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
See also \code{\link{subset}} and \code{\link[dplyr]{filter}} in \pkg{dplyr}.
}
\keyword{misc}

