% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fishR}
\alias{fishR}
\title{Opens web pages associated with the fishR website.}
\usage{
fishR(where = c("home", "general", "books", "AIFFD", "posts", "news"))
}
\arguments{
\item{where}{A string that indicates a particular page on the fishR website to open.}
}
\value{
None, but a webpage will be opened in the default browswer.
}
\description{
Opens web pages associated with the \href{http://fishr.wordpress.com"}{fishR website} in a browser.  The user can open the main page or choose a specific page to open.
}
\examples{
## ONLY RUN IN INTERACTIVE MODE
if (interactive()) {

fishR()            # home page
fishR("general")   # general vignettes page
fishR("books")     # books vignettes page
fishR("AIFFD")     # Analysis & Interpretation of Freshwater Fisheries Data page
fishR("posts")     # blog posts (some examples) page

} ## END IF INTERACTIVE MODE
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{misc}

