% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSA-internals.R
\name{.onAttach}
\alias{.onAttach}
\alias{iAddLoessLine}
\alias{iCheckALK}
\alias{iCheckStartcatW}
\alias{iCILabel}
\alias{iGetDecimals}
\alias{iGetVarFromFormula}
\alias{iHndlCols2UseIgnore}
\alias{iHndlFormula}
\alias{iHndlMultWhat}
\alias{iLegendHelp}
\alias{iListSpecies}
\alias{iMakeColor}
\alias{iRichColors}
\alias{iPlotExists}
\alias{is.wholenumber}
\alias{iTypeoflm}
\alias{STOP}
\alias{WARN}
\title{Internal functions.}
\usage{
.onAttach(lib, pkg, ...)
}
\description{
Internal functions that are common to several functions in FSA.
}
\keyword{internal}
