\name{fsaNews}
\alias{FSANews}
\alias{fsaNews}
\title{Read news and changes in the 'FSA' package.}
\usage{
fsaNews(...)
}
\arguments{
  \item{\dots}{Additional arguments passed to
  \code{file.show()}.}
}
\value{
None.
}
\description{
Read news and changes in the \sQuote{FSA} package.
}
\details{
This makes use of \code{file.show()} to display the
\sQuote{NEWS} file for this package.
}
\note{
Borrowed from Yihui Xie of the \sQuote{animation} package.
}
\examples{
fsaNews()
FSANews()
}
\keyword{manip}

