% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fsaNews}
\alias{FSANews}
\alias{fsaNews}
\title{Read news and changes for the 'FSA' package.}
\usage{
fsaNews()

FSANews()
}
\value{
None.
}
\description{
Opens up the \href{https://github.com/droglenc/FSA/blob/master/NEWS.md}{News.md GitHub file} for the \sQuote{FSA} package in an external browser.
}
\examples{
## Not run because it will open an external browser.
\dontrun{
fsaNews()
FSANews()
}
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{manip}

