% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsaNews.R
\name{fsaNews}
\alias{FSANews}
\alias{fsaNews}
\title{Read news and changes for the 'FSA' package.}
\usage{
fsaNews()

FSANews()
}
\value{
None.
}
\description{
Opens up the \href{https://github.com/droglenc/FSA/blob/master/NEWS.md}{News.md GitHub file} for the \sQuote{FSA} package in an external browser.
}
\examples{
## ONLY RUN IN INTERACTIVE MODE
if (interactive()) {

fsaNews()
FSANews()

}  ## END IF INTERACTIVE MODE
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\keyword{manip}

