% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSAUtils.R
\name{fsaNews}
\alias{fsaNews}
\alias{FSANews}
\title{Read news and changes for the 'FSA' package.}
\usage{
fsaNews()

FSANews()
}
\value{
None.
}
\description{
Opens up the \href{https://github.com/droglenc/FSA/blob/master/NEWS.md}{News.md GitHub file} for the \sQuote{FSA} package in an external browser.
}
\examples{
\dontrun{
## Opens an external webpage ... only run interactively
FSANews()
}

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
\keyword{manip}
