% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvalues.R
\name{mapvalues}
\alias{mapvalues}
\title{Replace specified values in a vector or factor with new values.}
\description{
Same function as in \pkg{plyr}.  Imported by \pkg{FSA} and exported to minimize conflicts with other commonly used packages.  See \code{\link[plyr]{mapvalues}} for documentation.
}
