% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdAdd}
\alias{psdAdd}
\alias{psdAdd.default}
\alias{psdAdd.formula}
\title{Create vector of Gabelhouse lengths for each species from an entire data frame.}
\usage{
psdAdd(len, ...)

\method{psdAdd}{default}(len, spec, units = c("mm", "cm", "in"),
  use.names = TRUE, ...)

\method{psdAdd}{formula}(len, data = NULL, units = c("mm", "cm", "in"),
  use.names = TRUE, ...)
}
\arguments{
\item{len}{A numeric vector that contains lengths measurements or a formula of the form \code{len~spec} where \dQuote{len} generically represents the length variable and \dQuote{spec} generically represents the species variable.  Note that this formula can only contain two variables and must have the length variable on the left-hand-side and the species variable on the right-hand-side.}

\item{spec}{A character or factor vector that contains the specicas names.  Not used if \code{len} is a formula.}

\item{units}{A string that indicates the type of units used for the lengths.  Choices are \code{mm} for millimeters (DEFAULT), \code{cm} for centimeters, and \code{in} for inches.}

\item{use.names}{A logical that indicates whether the vector returned is numeric (\code{=FALSE}) or word (\code{=TRUE}; default) representations of the Gabelhouse lengths.  See details.}

\item{data}{A data.frame that minimally contains the length measurements and species names if \code{len} is a formula.}

\item{\dots}{Not used.}
}
\value{
A numeric or factor vector that contains the Gabelhous length categories.
}
\description{
Creates a vector of the Gabelhouse lengths specific to a species for all individuals in an entire data frame.
}
\details{
This computes a vector that contains the Gabelhouse lengths specific to each species for all individuals in an entire data frame.  The vector can be appended to an existing data.frame to create a variable that contains the Gabelhous lengths for each individual.  The Gabelhouse length value will be \code{NA} for each individual for which a Gabelhouse length definitision do not exist in \code{\link{PSDlit}}.  Individuals shorter than \dQuote{stock} length will be listed as \code{zero} if \code{use.names=TRUE} or \code{0} if \code{use.names=FALSE}.  See the examples for one method for changing species names to something that this function will recognize.
}
\section{fishR vignette}{
 none yet.
}
\examples{
## Create random data for three species
# only for repeatability
set.seed(345234534)
dbg <- data.frame(species=factor(rep(c("Bluegill"),30)),tl=round(rnorm(30,130,50),0))
dbg$wt <- round(4.23e-06*dbg$tl^3.316+rnorm(30,0,10),1)
dlb <- data.frame(species=factor(rep(c("LMB"),30)),tl=round(rnorm(30,350,60),0))
dlb$wt <- round(2.96e-06*dlb$tl^3.273+rnorm(30,0,60),1)
dbt <- data.frame(species=factor(rep(c("bluefin tuna"),30)),tl=round(rnorm(30,1900,300),0))
dbt$wt <- round(4.5e-05*dbt$tl^2.8+rnorm(30,0,6000),1)
df <- rbind(dbg,dlb,dbt)
str(df)

# Rename variables to match PSDlit -- see recodeF()
df$species <- recodeF(df$species,"LMB","Largemouth Bass")
levels(df$species)

## Add variable using category names
# formula notation
df$PSD <- psdAdd(tl~species,data=df)
# non-formula notation
df$PSD1 <- psdAdd(df$tl,df$species)
# Add variable using length values as names
df$PSD2 <- psdAdd(tl~species,data=df,use.names=FALSE)

## All of the above but using dplyr
if (require(dplyr)) {
  df <- df \%>\%
    mutate(PSD1A=psdAdd(tl,species)) \%>\%
    mutate(PSD2A=psdAdd(tl,species,use.names=FALSE))
}

df
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
Guy, C.S., R.M. Neumann, and D.W. Willis.  2006.  \href{http://pubstorage.sdstate.edu/wfs/415-F.pdf}{New terminology for proportional stock density (PSD) and relative stock density (RSD): proportional size structure (PSS).}  Fisheries 31:86-87.

Guy, C.S., R.M. Neumann, D.W. Willis, and R.O. Anderson.  2006.  \href{http://www.montana.edu/mtcfru/Guy/Publication\%20pdf/PSD\%20pub.pdf}{Proportional size distribution (PSD): A further refinement of population size structure index terminology.}  Fisheries 32:348.

Willis, D.W., B.R. Murphy, and C.S. Guy.  1993.  \href{http://web1.cnre.vt.edu/murphybr/web/Readings/Willis\%20et\%20al.pdf}{Stock density indices: development, use, and limitations.}  Reviews in Fisheries Science 1:203-222.
}
\seealso{
\code{\link{psdVal}}, \code{\link{psdCalc}}, \code{\link{psdPlot}}, \code{\link{PSDlit}}, \code{\link{recodeF}}, and \code{\link{wrAdd}}.
}
\keyword{manip}

