% Generated by roxygen2 (4.0.1): do not edit by hand
\name{psdCalc}
\alias{psdCalc}
\title{Convenience function for calculating PSD-X and PSD X-Y values.}
\usage{
psdCalc(formula, data, species = "List", units = c("mm", "cm", "in"),
  what = c("all", "traditional", "interval", "none"), method = c("multinom",
  "binom", "Gustafson"), conf.level = 0.95, addLens = NULL,
  addNames = NULL, digits = getOption("digits"))
}
\arguments{
\item{formula}{A formula of the form \code{~length} where \dQuote{length} generically represents a variable in \code{data} that contains length measurements.  Note that this formula may only contain one variable.}

\item{data}{A data.frame that minimally contains the length measurements given in the variable in the \code{formula}.}

\item{species}{A string that contains the species name for which Gabelhouse length categories exist.  See \code{\link{psdVal}} for details.}

\item{units}{A string that indicates the type of units used for the length measurements.  Choices are \code{mm} for millimeters (DEFAULT), \code{cm} for centimeters, and \code{in} for inches.}

\item{what}{A string that indicates the level of PSD values that will be printed.  See details.}

\item{method}{A character that identifies the confidence interval method to use.  See details in \code{\link{psdCI}}.}

\item{addLens}{A numeric vector that contains minimum length definitions for additional categories.  See \code{\link{psdVal}} for details.}

\item{addNames}{A string vector that contains names for the additional length categories added with \code{addLens}.  See \code{\link{psdVal}} for details.}

\item{conf.level}{A number that indicates the level of confidence to use for constructing confidence intervals (default is \code{0.95}).}

\item{digits}{A numeric that indicates the number of decimals to round the result to.}
}
\value{
A matrix with columns that contain the computed PSD-X or PSD X-Y values and the associated confidence intervals.
}
\description{
Convenience function for calculating (traditional) PSD-X and (interval) PSD X-Y values for all lengths and intervals of Gabelhouse lengths.
}
\details{
This computes the  (traditional) PSD-X and (interval) PSD X-Y values, with associated confidence intervals, for each Gabelhouse length.  All PSD-X and PSD X-Y values are printed if \code{what="all"} (DEFAULT), only PSD-X values are printed if \code{what="traditional"}, only PSD X-Y values are printed if \code{what="interval"}, and nothing is printed (but the matrix is still returned) if \code{what="none"}.
}
\section{fishR vignette}{
 \url{https://sites.google.com/site/fishrfiles/gnrl/SizeStructure.pdf}
}
\examples{
## Random length data
# suppose this is yellow perch to the nearest mm
yepdf <- data.frame(yepmm=c(rnorm(100,mean=125,sd=15),rnorm(50,mean=200,sd=25),
                    rnorm(20,mean=300,sd=40)),
                    species=rep("Yellow Perch",170))
psdCalc(~yepmm,data=yepdf,species="Yellow perch",digits=1)

## all values above stock value (troubleshooting a problem)
yepdf2 <- subset(yepdf,yepmm>=130)
psdCalc(~yepmm,data=yepdf2,species="Yellow perch",digits=1)

## all values above quality value (troubleshooting a problem)
yepdf3 <- subset(yepdf,yepmm>=200)
psdCalc(~yepmm,data=yepdf3,species="Yellow perch",digits=1)

## add a length
psdCalc(~yepmm,data=yepdf,species="Yellow perch",addLens=150,digits=1)

## add a length with a name
psdCalc(~yepmm,data=yepdf,species="Yellow perch",addLens=150,addNames="minLen",digits=1)

## different output types
psdCalc(~yepmm,data=yepdf,species="Yellow perch",addLens=150,digits=1,what="traditional")
psdCalc(~yepmm,data=yepdf,species="Yellow perch",addLens=150,digits=1,what="interval")
psdCalc(~yepmm,data=yepdf,species="Yellow perch",addLens=150,digits=1,what="none")
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\references{
Guy, C.S., R.M. Neumann, and D.W. Willis.  2006.  \href{http://pubstorage.sdstate.edu/wfs/415-F.pdf}{New terminology for proportional stock density (PSD) and relative stock density (RSD): proportional size structure (PSS).}  Fisheries 31:86-87.

Guy, C.S., R.M. Neumann, D.W. Willis, and R.O. Anderson.  2006.  \href{http://www.montana.edu/mtcfru/Guy/Publication\%20pdf/PSD\%20pub.pdf}{Proportional size distribution (PSD): A further refinement of population size structure index terminology.}  Fisheries 32:348.

Willis, D.W., B.R. Murphy, and C.S. Guy.  1993.  \href{http://web1.cnre.vt.edu/murphybr/web/Readings/Willis\%20et\%20al.pdf}{Stock density indices: development, use, and limitations.}  Reviews in Fisheries Science 1:203-222.
}
\seealso{
\code{\link{psdVal}}, \code{\link{psdPlot}}, \code{\link{psdDataPrep}}, \code{\link{PSDlit}}, \code{\link{tictactoe}}, \code{\link{tictactoeAdd}}, \code{\link{lencat}}, and \code{\link{rcumsum}}.
}
\keyword{hplot}

