% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGrowthFun.R
\name{showGrowthFun}
\alias{showGrowthFun}
\title{Creates an expression for a specific parameterization of the von Bertalanffy, Gompertz, Richards, and logistic growth functions, as well as the Schnute and Schnute-Richards growth functions.}
\usage{
showGrowthFun(
  type = c("von Bertalanffy", "Gompertz", "Richards", "logistic", "Schnute",
    "Schnute-Richards"),
  param = 1,
  case = NULL,
  plot = FALSE,
  ...
)
}
\arguments{
\item{type}{A string (\dQuote{von Bertalanffy}, \dQuote{Gompertz}, \dQuote{logistic}, \dQuote{Richards} \dQuote{Schnute-Richards}) that indicates the type of growth function to show.}

\item{param}{A single numeric that indicates the specific parameterization of the growth function. See details.}

\item{case}{A numeric that indicates the specific case of the Schnute function to use. See details.}

\item{plot}{A logical for whether the expression should be shown on a \dQuote{blank} plot. See examples.}

\item{\dots}{Arguments for \code{plot}. In particular use \code{cex=} to make the expression larger and easier to read. See examples.}
}
\value{
An expression representing the equation of the growth function given in \code{type} and \code{param}.
}
\description{
Creates an expression for a specific parameterization of the von Bertalanffy, Gompertz, Richards, and logistic growth functions, as well as the Schnute and Schnute-Richards growth functions. The expression can be plotted with \code{plot=TRUE} to see the equation of the growth function.
}
\examples{
showGrowthFun()
showGrowthFun(plot=TRUE)
showGrowthFun(plot=TRUE,cex=2)
showGrowthFun(type="Richards",param=3,plot=TRUE,cex=1.5)

}
\seealso{
See \code{\link{makeGrowthFun}} to make functions that correspond to these expressions.
}
