% Generated by roxygen2 (4.0.0): do not edit by hand
\name{srModels}
\alias{srModels}
\title{Show the stock-recruitment model formulas implemented in FSA.}
\usage{
srModels(...)
}
\arguments{
\item{\dots}{Additional arguments for \code{plot}.  Generally not needed}
}
\value{
A graphic that uses \code{plotmath} to show the model formulas in a
pretty format.
}
\description{
Show the stock-recruitment model formulas implemented in \code{\link{srFuns}},
\code{\link{srStarts}}, and \code{srSim} in \pkg{FSATeach}.
}
\section{fishR vignette}{
 \url{https://sites.google.com/site/fishrfiles/gnrl/StockRecruit.pdf}
}
\examples{
srModels()
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
\code{\link{srFuns}}, \code{\link{srStarts}}, and \code{\link{srSim}}
}
\keyword{hplot}
\keyword{manip}

