% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validn.R
\name{validn}
\alias{validn}
\title{Finds the number of valid (non-NA) values in a vector.}
\usage{
validn(object)
}
\arguments{
\item{object}{A vector.}
}
\value{
A single numeric value that is the number of non-\code{NA} values in a vector.
}
\description{
Finds the number of valid (non-NA) values in a vector.
}
\section{IFAR Chapter}{
 \href{https://fishr.wordpress.com/books/ifar/}{2-Basic Data Manipulations}.
}
\examples{
junk1 <- c(1,7,2,4,3,10,NA)
junk2 <- c("Derek","Hugh","Ogle","Santa","Claus","Nick",NA,NA)
junk3 <- factor(junk2)
junk4 <- c(TRUE,TRUE,FALSE,FALSE,FALSE,TRUE,NA,NA)

validn(junk1)
validn(junk2)
validn(junk3)
validn(junk4)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
See \code{\link[plotrix]{valid.n}} in \pkg{plotrix} and \code{\link[gdata]{nobs}} in \pkg{gdata} for similar functionality.  See \code{\link{is.na}} for finding the missing values.
}
\keyword{manip}

