% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/walfordChapmanPlot.R
\name{walfordPlot}
\alias{chapmanPlot}
\alias{chapmanPlot.default}
\alias{chapmanPlot.formula}
\alias{walfordPlot}
\alias{walfordPlot.default}
\alias{walfordPlot.formula}
\title{Constructs Ford-Walford and Chapman plots for mean length-at-age data.}
\usage{
walfordPlot(formula, data = NULL, xlab = expression(L[t]),
  ylab = expression(L[t + 1]), colLS = "blue", colRL = "red", lwdLS = 2,
  lwdRL = 2, ltyLS = 1, ltyRL = 2, pch = 19, col = "black",
  pchLinf = 9, colLinf = "red", cexLinf = 1.5, showLS = TRUE,
  showVB = TRUE, ...)

chapmanPlot(formula, data = NULL, xlab = expression(L[t]),
  ylab = expression(L[t + 1] - L[t]), colLS = "blue", ltyLS = 1,
  lwdLS = 2, pch = 19, col = "black", pchLinf = 9, colLinf = "red",
  cexLinf = 1.5, showLS = TRUE, showVB = TRUE, ...)
}
\arguments{
\item{formula}{A formula of the form \code{len~age}.}

\item{data}{A data frame in which \code{age} and \code{len} can be found.}

\item{xlab}{A string for labeling the x-axis.}

\item{ylab}{A string for labeling the y-axis.}

\item{colLS}{A string that indicates the color of the least-squares line.}

\item{colRL}{A string that indicates the color of the 1:1 reference line.}

\item{lwdLS}{A numeric that indicates the line width of the least-squares line.}

\item{lwdRL}{A numeric that indicates the line width of the 1:1 reference line.}

\item{ltyLS}{A numeric that indicates the line type of the least-squares line.}

\item{ltyRL}{A numeric that indicates the line type of the 1:1 reference line.}

\item{pch}{A numeric that indicates the plotting character type for the observed data.}

\item{col}{A string that indicates the color for the observed data.}

\item{pchLinf}{A numeric that indicates the plotting character type that indicates the Linf value.}

\item{colLinf}{A string that indicates the color of the point that indicates the Linf value.}

\item{cexLinf}{A numeric that indicates the character expansion factor of the point that indicates the Linf value.}

\item{showLS}{A logical that indicates whether the coefficients from the least-squares fit should be shown on the plot.}

\item{showVB}{A logical that indicates whether the estimate of Linf and K should be shown on the plot.}

\item{\dots}{Other arguments to be passed to the \code{plot} function.}
}
\value{
None.  However, a plot is produced.
}
\description{
Constructs Ford-Walford and Chapman plots for mean length-at-age data and computes estimates of asymptotic mean length (Linf) and the Brody growth coefficient (K).
}
\examples{
data(SMBassWB)
walfordPlot(lencap~agecap,data=SMBassWB)
chapmanPlot(lencap~agecap,data=SMBassWB)
}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
\code{\link{vbStarts}}
}
\keyword{hplot}

